[asm] Modularize TranslateFromMLIR with operation handlers
Major refactoring of TranslateFromMLIR.cpp to improve maintainability:

- Add modular handler architecture with separate files for each dialect:
  - GPUHandlers.cpp: GPU dialect operations
  - ArithHandlers.cpp: Arithmetic operations
  - MathHandlers.cpp: Math dialect operations
  - VectorHandlers.cpp: Vector operations
  - SCFHandlers.cpp: SCF control flow (for, if, yield)
  - MemRefHandlers.cpp: Memory operations (alloc, load, store, etc.)
  - AffineHandlers.cpp: Affine apply with optimizations
  - AMDGPUHandlers.cpp: AMDGPU/ROCDL operations (mfma, buffer ops, g2s)

- Add shared infrastructure:
  - Handlers.h: Common declarations and helper functions
  - ABI.h: AMDGCN ABI constants (VGPR/SGPR counts, stack alignment)
  - Helper templates for common patterns (binary ops, unary ops)

- Add operation verifiers for safety checks
- Fix logic issues and add safety checks throughout
- Remove unused helper functions to eliminate compilation warnings

Reduces TranslateFromMLIR.cpp from ~4600 lines to ~1900 lines while
maintaining all functionality and passing all tests.
