+++ wave_lang/kernel/wave/asm/wave_asm/lib/Transforms/TranslateFromMLIR.cpp
@@ -310,6 +310,66 @@ void TranslationContext::updateSRDBufferSize(Value memref, int64_t bufferSize) {
 using namespace waveasm;
 
 //===----------------------------------------------------------------------===//
+// Forward Declarations for External Handler Files
+//===----------------------------------------------------------------------===//
+// These handlers are implemented in separate files under handlers/
+// Forward declarations allow registerDefaultHandlers to find them.
+
+namespace waveasm {
+// GPU handlers (implemented in handlers/GPUHandlers.cpp)
+LogicalResult handleGPUThreadId(Operation *op, TranslationContext &ctx);
+LogicalResult handleGPUBlockId(Operation *op, TranslationContext &ctx);
+LogicalResult handleGPUBarrier(Operation *op, TranslationContext &ctx);
+LogicalResult handleGPUBlockDim(Operation *op, TranslationContext &ctx);
+LogicalResult handleGPUGridDim(Operation *op, TranslationContext &ctx);
+LogicalResult handleGPULaneId(Operation *op, TranslationContext &ctx);
+LogicalResult handleGPUSubgroupBroadcast(Operation *op,
+                                         TranslationContext &ctx);
+
+// Arith handlers (implemented in handlers/ArithHandlers.cpp)
+LogicalResult handleArithConstant(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithAddI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithSubI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithMulI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithDivUI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithRemUI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithIndexCast(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithAndI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithOrI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithXorI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithShLI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithShRUI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithShRSI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithExtUI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithExtSI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithTruncI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithCmpI(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithSelect(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithAddF(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithSubF(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithMulF(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithDivF(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithNegF(Operation *op, TranslationContext &ctx);
+LogicalResult handleArithCmpF(Operation *op, TranslationContext &ctx);
+
+// Math handlers (implemented in handlers/MathHandlers.cpp)
+LogicalResult handleMathFma(Operation *op, TranslationContext &ctx);
+
+// Vector handlers (implemented in handlers/VectorHandlers.cpp)
+// Note: handleVectorLoad, handleVectorStore, handleVectorTransferRead,
+// handleVectorTransferWrite remain in this file due to their complexity.
+LogicalResult handleVectorBroadcast(Operation *op, TranslationContext &ctx);
+LogicalResult handleVectorExtract(Operation *op, TranslationContext &ctx);
+LogicalResult handleVectorInsert(Operation *op, TranslationContext &ctx);
+LogicalResult handleVectorShapeCast(Operation *op, TranslationContext &ctx);
+LogicalResult handleVectorFma(Operation *op, TranslationContext &ctx);
+LogicalResult handleVectorReduction(Operation *op, TranslationContext &ctx);
+LogicalResult handleVectorExtractStridedSlice(Operation *op,
+                                              TranslationContext &ctx);
+
+} // namespace waveasm
+
+//===----------------------------------------------------------------------===//
 // Helper Functions
 //===----------------------------------------------------------------------===//
 
@@ -327,6 +387,37 @@ static int64_t computeBufferSizeFromMemRef(MemRefType memrefType) {
   if (elementBytes == 0)
     elementBytes = 1; // Minimum 1 byte
   return numElements * elementBytes;
+}
+
+/// Check if a memref has LDS address space
+bool isLDSMemRef(MemRefType memrefType) {
+  auto memSpace = memrefType.getMemorySpace();
+  if (!memSpace)
+    return false;
+
+  // Check for gpu.address_space<workgroup> attribute
+  if (auto gpuSpace = dyn_cast<gpu::AddressSpaceAttr>(memSpace)) {
+    return gpuSpace.getValue() == gpu::AddressSpace::Workgroup;
+  }
+  // Also check for integer address space (3 on AMDGPU)
+  if (auto intAttr = dyn_cast<IntegerAttr>(memSpace)) {
+    return intAttr.getInt() == 3;
+  }
+  return false;
+}
+
+/// Get element size in bytes
+int64_t getElementBytes(Type type) {
+  if (auto floatType = dyn_cast<FloatType>(type))
+    return floatType.getWidth() / 8;
+  if (auto intType = dyn_cast<IntegerType>(type))
+    return (intType.getWidth() + 7) / 8;
+  return 4;
+}
+
+/// Get vector size in bytes
+int64_t getVectorBytes(VectorType vecType) {
+  return vecType.getNumElements() * getElementBytes(vecType.getElementType());
 }
 
 /// Get or create a mapped VGPR value for an MLIR value
@@ -359,1137 +450,14 @@ static Value getOrCreateVReg(Value mlirValue, TranslationContext &ctx,
   return mlirValue; // Return as-is for now, handler will create proper op
 }
 
-/// Check if a memref has LDS address space
-bool isLDSMemRef(MemRefType memrefType) {
-  auto memSpace = memrefType.getMemorySpace();
-  if (!memSpace)
-    return false;
-
-  if (auto gpuSpace = dyn_cast<gpu::AddressSpaceAttr>(memSpace)) {
-    return gpuSpace.getValue() == gpu::AddressSpace::Workgroup;
-  }
-  return false;
-}
-
-/// Get element size in bytes
-int64_t getElementBytes(Type type) {
-  if (auto floatType = dyn_cast<FloatType>(type))
-    return floatType.getWidth() / 8;
-  if (auto intType = dyn_cast<IntegerType>(type))
-    return (intType.getWidth() + 7) / 8;
-  return 4;
-}
-
-/// Get vector size in bytes
-int64_t getVectorBytes(VectorType vecType) {
-  return vecType.getNumElements() * getElementBytes(vecType.getElementType());
-}
-
-/// Check if value is power of 2
-bool isPowerOf2(int64_t val) { return val > 0 && (val & (val - 1)) == 0; }
-
-/// Get log2 of power of 2
-int64_t log2(int64_t val) {
-  int64_t result = 0;
-  while (val > 1) {
-    val >>= 1;
-    result++;
-  }
-  return result;
-}
-
-//===----------------------------------------------------------------------===//
-// GPU Dialect Handlers
-//===----------------------------------------------------------------------===//
-
-/// Handle gpu.thread_id - emit v_mbcnt for single-wave or use v_bfe_u32 for
-/// multi-wave
-LogicalResult handleGPUThreadId(Operation *op, TranslationContext &ctx) {
-  auto threadIdOp = cast<gpu::ThreadIdOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  Value result;
-
-  // Get the dimension being requested
-  gpu::Dimension dim = threadIdOp.getDimension();
-
-  // Check if this is a multi-wave kernel
-  // For multi-wave, the hardware provides flat_workitem_id in v0
-  // For single-wave, we compute lane_id using v_mbcnt
-  if (ctx.isMultiWaveKernel()) {
-    // Multi-wave: hardware provides flat workitem ID in v0
-    // The flat workitem ID is packed as: tid_x + tid_y * wg_x + tid_z * wg_x *
-    // wg_y We use v_bfe_u32 to extract the individual components Following
-    // Python's approach: 10 bits per dimension (supports up to 1024 per dim)
-
-    // Mark that this kernel uses workitem ID (set
-    // amdhsa_system_vgpr_workitem_id)
-    ctx.setUsesWorkitemId(true);
-
-    // Get flat workitem ID from v0
-    auto flatWorkitemId =
-        PrecoloredVRegOp::create(builder, loc, vregType, 0, 1);
-
-    // Determine bit offset and width based on dimension
-    // Python uses 10 bits per dimension:
-    // - tid_x: bits 0-9 (offset=0, width=10)
-    // - tid_y: bits 10-19 (offset=10, width=10)
-    // - tid_z: bits 20-29 (offset=20, width=10)
-    int64_t bitOffset = 0;
-    int64_t bitWidth = 10;
-
-    switch (dim) {
-    case gpu::Dimension::x:
-      bitOffset = 0;
-      break;
-    case gpu::Dimension::y:
-      bitOffset = 10;
-      break;
-    case gpu::Dimension::z:
-      bitOffset = 20;
-      break;
-    }
-
-    // Create constants for offset and width
-    auto immOffset = ctx.createImmType(bitOffset);
-    auto offsetConst = ConstantOp::create(builder, loc, immOffset, bitOffset);
-
-    auto immWidth = ctx.createImmType(bitWidth);
-    auto widthConst = ConstantOp::create(builder, loc, immWidth, bitWidth);
-
-    // v_bfe_u32 dst, src, offset, width - extract bits [offset, offset+width-1]
-    result = V_BFE_U32::create(builder, loc, vregType, flatWorkitemId,
-                               offsetConst, widthConst);
-  } else {
-    // Single-wave: compute lane ID using v_mbcnt
-    // Note: for single-wave, tid_x == lane_id, and tid_y/tid_z are always 0
-
-    if (dim == gpu::Dimension::x) {
-      // Create constant -1 for the mask
-      auto immType = ctx.createImmType(-1);
-      auto maskConst = ConstantOp::create(builder, loc, immType, -1);
-
-      // Create constant 0
-      auto immZero = ctx.createImmType(0);
-      auto zeroConst = ConstantOp::create(builder, loc, immZero, 0);
-
-      // v_mbcnt_lo_u32_b32 - count bits in lower 32 lanes
-      auto mbcntLo = V_MBCNT_LO_U32_B32::create(builder, loc, vregType,
-                                                maskConst, zeroConst);
-
-      // v_mbcnt_hi_u32_b32 - count bits in upper 32 lanes, add to low
-      auto mbcntHi = V_MBCNT_HI_U32_B32::create(builder, loc, vregType,
-                                                maskConst, mbcntLo);
-
-      result = mbcntHi;
-    } else {
-      // For single-wave, tid_y and tid_z are always 0
-      auto immZero = ctx.createImmType(0);
-      result = ConstantOp::create(builder, loc, immZero, 0);
-    }
-
-    // Note: for single-wave, we don't set usesWorkitemId since we use v_mbcnt
-    // The kernel descriptor should have amdhsa_system_vgpr_workitem_id = 0
-  }
-
-  // Map the result
-  ctx.getMapper().mapValue(threadIdOp.getResult(), result);
-
-  // Track upper bound for affine simplification
-  // The upper_bound attribute tells us the range of this thread ID
-  if (auto upperBoundAttr = threadIdOp.getUpperBoundAttr()) {
-    int64_t upperBound = upperBoundAttr.getInt();
-    // Track this thread ID value with its upper bound
-    ctx.setThreadIdUpperBound(threadIdOp.getResult(), upperBound);
-
-    // Also set bit range for OR vs ADD optimization
-    // Thread ID is in range [0, upperBound-1], so we need bits
-    // 0..log2(upperBound-1)
-    ctx.setBitRange(result, BitRange::fromMaxValue(upperBound - 1));
-  } else {
-    // Default for 64-lane wavefront: [0, 63] needs 6 bits
-    ctx.setBitRange(result, BitRange(0, 5));
-  }
-
-  return success();
-}
-
-/// Handle gpu.block_id - block ID comes from system SGPRs
-LogicalResult handleGPUBlockId(Operation *op, TranslationContext &ctx) {
-  auto blockIdOp = cast<gpu::BlockIdOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Block IDs are passed in system SGPRs via ABI
-  // System SGPRs come after user SGPRs (kernarg ptr + preloaded args)
-  gpu::Dimension dim = blockIdOp.getDimension();
-  int dimIndex = 0;
-  switch (dim) {
-  case gpu::Dimension::x:
-    dimIndex = 0;
-    ctx.setUsesWorkgroupIdX(true);
-    break;
-  case gpu::Dimension::y:
-    dimIndex = 1;
-    ctx.setUsesWorkgroupIdY(true);
-    break;
-  case gpu::Dimension::z:
-    dimIndex = 2;
-    ctx.setUsesWorkgroupIdZ(true);
-    break;
-  }
-
-  int64_t sgprIndex = ctx.getWorkgroupIdSgprIndex(dimIndex);
-
-  auto sregType = ctx.createSRegType();
-  auto blockId = PrecoloredSRegOp::create(builder, loc, sregType, sgprIndex, 1);
-
-  ctx.getMapper().mapValue(blockIdOp.getResult(), blockId);
-  return success();
-}
-
-/// Handle gpu.barrier - emit s_barrier
-LogicalResult handleGPUBarrier(Operation *op, TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  S_BARRIER::create(builder, loc);
-  return success();
-}
-
-/// Handle gpu.block_dim - workgroup dimensions from kernel dispatch packet
-LogicalResult handleGPUBlockDim(Operation *op, TranslationContext &ctx) {
-  auto blockDimOp = cast<gpu::BlockDimOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Block dimensions are available from the dispatch packet or as kernel args
-  // For now, use precolored SGPRs at standard ABI locations
-  int64_t sgprIndex = 0;
-  switch (blockDimOp.getDimension()) {
-  case gpu::Dimension::x:
-    sgprIndex = 6;
-    break; // workgroup_size_x
-  case gpu::Dimension::y:
-    sgprIndex = 7;
-    break; // workgroup_size_y
-  case gpu::Dimension::z:
-    sgprIndex = 8;
-    break; // workgroup_size_z
-  }
-
-  auto sregType = ctx.createSRegType();
-  auto blockDim =
-      PrecoloredSRegOp::create(builder, loc, sregType, sgprIndex, 1);
-
-  ctx.getMapper().mapValue(blockDimOp.getResult(), blockDim);
-  return success();
-}
-
-/// Handle gpu.grid_dim - grid dimensions from kernel dispatch packet
-LogicalResult handleGPUGridDim(Operation *op, TranslationContext &ctx) {
-  auto gridDimOp = cast<gpu::GridDimOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Grid dimensions come from dispatch packet
-  int64_t sgprIndex = 0;
-  switch (gridDimOp.getDimension()) {
-  case gpu::Dimension::x:
-    sgprIndex = 9;
-    break; // grid_size_x
-  case gpu::Dimension::y:
-    sgprIndex = 10;
-    break; // grid_size_y
-  case gpu::Dimension::z:
-    sgprIndex = 11;
-    break; // grid_size_z
-  }
-
-  auto sregType = ctx.createSRegType();
-  auto gridDim = PrecoloredSRegOp::create(builder, loc, sregType, sgprIndex, 1);
-
-  ctx.getMapper().mapValue(gridDimOp.getResult(), gridDim);
-  return success();
-}
-
-/// Handle gpu.lane_id - lane within wavefront (same as thread_id for wave64)
-LogicalResult handleGPULaneId(Operation *op, TranslationContext &ctx) {
-  auto laneIdOp = cast<gpu::LaneIdOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  // Lane ID is the thread's position within the wavefront
-  // Use v_mbcnt_lo + v_mbcnt_hi pattern
-  auto immType = ctx.createImmType(-1);
-  auto maskConst = ConstantOp::create(builder, loc, immType, -1);
-
-  auto immZero = ctx.createImmType(0);
-  auto zeroConst = ConstantOp::create(builder, loc, immZero, 0);
-
-  auto mbcntLo =
-      V_MBCNT_LO_U32_B32::create(builder, loc, vregType, maskConst, zeroConst);
-  auto mbcntHi =
-      V_MBCNT_HI_U32_B32::create(builder, loc, vregType, maskConst, mbcntLo);
-
-  ctx.getMapper().mapValue(laneIdOp.getResult(), mbcntHi);
-  return success();
-}
-
-/// Handle gpu.subgroup_broadcast - broadcast value from one lane to all
-/// This emits v_readlane_b32 to read from a specific lane, or
-/// v_readfirstlane_b32 for broadcasting from lane 0.
-LogicalResult handleGPUSubgroupBroadcast(Operation *op,
-                                         TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // gpu.subgroup_broadcast can have either:
-  // - 2 operands: value and lane_id
-  // - 1 operand with lane_id as attribute (newer MLIR versions)
-  if (op->getNumOperands() < 1) {
-    return op->emitError(
-        "subgroup_broadcast requires at least a value operand");
-  }
-
-  Value srcValue = op->getOperand(0);
-  auto srcMapped = ctx.getMapper().getMapped(srcValue);
-  if (!srcMapped) {
-    return op->emitError("source value not mapped");
-  }
-
-  // If the source is an immediate (constant), broadcasting is a no-op
-  // Just map the result to the same immediate value
-  if (isa<ImmType>(srcMapped->getType())) {
-    ctx.getMapper().mapValue(op->getResult(0), *srcMapped);
-    return success();
-  }
-
-  // Check for broadcast_type attribute first (newer MLIR with gpu.broadcast
-  // enum) #gpu<broadcast first_active_lane> means use v_readfirstlane_b32
-  bool usesFirstLane = false;
-  if (auto broadcastTypeAttr = op->getAttr("broadcast_type")) {
-    // The attribute is an enum like #gpu<broadcast first_active_lane>
-    // For gpu.subgroup_broadcast with broadcast_type = first_active_lane,
-    // we should use v_readfirstlane_b32
-    // The attribute prints as "#gpu<broadcast first_active_lane>"
-    // TODO: Use mlir::gpu::BroadcastTypeAttr when available to avoid string
-    // matching
-    std::string attrDump;
-    llvm::raw_string_ostream os(attrDump);
-    broadcastTypeAttr.print(os);
-    if (attrDump.find("first_active_lane") != std::string::npos) {
-      usesFirstLane = true;
-    }
-  }
-
-  // Check for lane_id - either as operand or attribute
-  std::optional<Value> laneMapped;
-  int64_t laneIdValue = 0;
-  bool hasLaneId = false;
-
-  if (op->getNumOperands() >= 2) {
-    // Lane ID is operand 1
-    Value laneIdx = op->getOperand(1);
-    laneMapped = ctx.getMapper().getMapped(laneIdx);
-    hasLaneId = true;
-  } else if (auto laneAttr = op->getAttrOfType<IntegerAttr>("lane_id")) {
-    // Lane ID is an attribute (newer MLIR)
-    laneIdValue = laneAttr.getInt();
-    hasLaneId = true;
-  } else if (auto subgroupIdAttr =
-                 op->getAttrOfType<IntegerAttr>("subgroup_id")) {
-    // Alternative attribute name
-    laneIdValue = subgroupIdAttr.getInt();
-    hasLaneId = true;
-  }
-
-  auto sregType = ctx.createSRegType();
-
-  // If lane is constant 0, also use v_readfirstlane_b32
-  if (!usesFirstLane) {
-    if (laneMapped) {
-      if (auto constOp = laneMapped->getDefiningOp<ConstantOp>()) {
-        if (constOp.getValue() == 0) {
-          usesFirstLane = true;
-        }
-      }
-    } else if (hasLaneId && laneIdValue == 0) {
-      usesFirstLane = true;
-    }
-  }
-
-  Value result;
-  if (usesFirstLane) {
-    result = V_READFIRSTLANE_B32::create(builder, loc, sregType, *srcMapped);
-  } else {
-    // Use v_readlane_b32 with lane index
-    if (!laneMapped) {
-      auto immType = ctx.createImmType(laneIdValue);
-      laneMapped = ConstantOp::create(builder, loc, immType, laneIdValue);
-    }
-    result =
-        V_READLANE_B32::create(builder, loc, sregType, *srcMapped, *laneMapped);
-  }
-
-  ctx.getMapper().mapValue(op->getResult(0), result);
-  return success();
-}
-
-//===----------------------------------------------------------------------===//
-// Arith Dialect Handlers
-//===----------------------------------------------------------------------===//
-
-/// Handle arith.constant - create immediate or track value
-LogicalResult handleArithConstant(Operation *op, TranslationContext &ctx) {
-  auto constOp = cast<arith::ConstantOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  auto value = constOp.getValue();
-
-  // Handle integer constants
-  if (auto intAttr = dyn_cast<IntegerAttr>(value)) {
-    int64_t intVal = intAttr.getInt();
-    auto immType = ctx.createImmType(intVal);
-    auto immOp = ConstantOp::create(builder, loc, immType, intVal);
-    ctx.getMapper().mapValue(constOp.getResult(), immOp);
-    return success();
-  }
-
-  // Handle float constants - need to bitcast to int
-  if (auto floatAttr = dyn_cast<FloatAttr>(value)) {
-    double floatVal = floatAttr.getValueAsDouble();
-    // For f32, bitcast to i32
-    if (floatAttr.getType().isF32()) {
-      float f = static_cast<float>(floatVal);
-      int32_t bits;
-      memcpy(&bits, &f, sizeof(bits));
-      auto immType = ctx.createImmType(bits);
-      auto immOp = ConstantOp::create(builder, loc, immType, bits);
-      ctx.getMapper().mapValue(constOp.getResult(), immOp);
-    }
-    return success();
-  }
-
-  // Handle dense vector constants (e.g., dense<0.0> for accumulator init)
-  if (auto denseAttr = dyn_cast<DenseElementsAttr>(value)) {
-    if (denseAttr.isSplat()) {
-      auto splatVal = denseAttr.getSplatValue<Attribute>();
-      if (auto floatAttr = dyn_cast<FloatAttr>(splatVal)) {
-        double floatVal = floatAttr.getValueAsDouble();
-        if (floatVal == 0.0) {
-          // Zero accumulator - use literal 0
-          auto immType = ctx.createImmType(0);
-          auto immOp = ConstantOp::create(builder, loc, immType, 0);
-          ctx.getMapper().mapValue(constOp.getResult(), immOp);
-        }
-      }
-    }
-    return success();
-  }
-
-  return success();
-}
-
-/// Handle arith.addi - emit v_add_u32
-LogicalResult handleArithAddI(Operation *op, TranslationContext &ctx) {
-  auto addOp = cast<arith::AddIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(addOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(addOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto add = V_ADD_U32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(addOp.getResult(), add);
-  return success();
-}
-
-/// Handle arith.muli - emit v_mul_lo_u32
-LogicalResult handleArithMulI(Operation *op, TranslationContext &ctx) {
-  auto mulOp = cast<arith::MulIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(mulOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(mulOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto mul = V_MUL_LO_U32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(mulOp.getResult(), mul);
-  return success();
-}
-
-/// Handle arith.index_cast - typically a no-op on GPU
-LogicalResult handleArithIndexCast(Operation *op, TranslationContext &ctx) {
-  auto castOp = cast<arith::IndexCastOp>(op);
-
-  // Index casts are usually no-ops (same register, different type
-  // interpretation)
-  auto src = ctx.getMapper().getMapped(castOp.getIn());
-  if (src) {
-    ctx.getMapper().mapValue(castOp.getResult(), *src);
-  }
-  return success();
-}
-
-/// Handle arith.subi - emit v_sub_u32
-LogicalResult handleArithSubI(Operation *op, TranslationContext &ctx) {
-  auto subOp = cast<arith::SubIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(subOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(subOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto sub = V_SUB_U32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(subOp.getResult(), sub);
-  return success();
-}
-
-/// Handle arith.andi - emit v_and_b32
-LogicalResult handleArithAndI(Operation *op, TranslationContext &ctx) {
-  auto andOp = cast<arith::AndIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(andOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(andOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto andResult = V_AND_B32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(andOp.getResult(), andResult);
-  return success();
-}
-
-/// Handle arith.ori - emit v_or_b32
-LogicalResult handleArithOrI(Operation *op, TranslationContext &ctx) {
-  auto orOp = cast<arith::OrIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(orOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(orOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto orResult = V_OR_B32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(orOp.getResult(), orResult);
-  return success();
-}
-
-/// Handle arith.xori - emit v_xor_b32
-LogicalResult handleArithXorI(Operation *op, TranslationContext &ctx) {
-  auto xorOp = cast<arith::XOrIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(xorOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(xorOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto xorResult = V_XOR_B32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(xorOp.getResult(), xorResult);
-  return success();
-}
-
-/// Handle arith.shli - emit v_lshlrev_b32
-LogicalResult handleArithShLI(Operation *op, TranslationContext &ctx) {
-  auto shlOp = cast<arith::ShLIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(shlOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(shlOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  // v_lshlrev has reversed operands (shift amount first)
-  auto shl = V_LSHLREV_B32::create(builder, loc, vregType, *rhs, *lhs);
-  ctx.getMapper().mapValue(shlOp.getResult(), shl);
-  return success();
-}
-
-/// Handle arith.shrui - emit v_lshrrev_b32 (unsigned right shift)
-LogicalResult handleArithShRUI(Operation *op, TranslationContext &ctx) {
-  auto shrOp = cast<arith::ShRUIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(shrOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(shrOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  // v_lshrrev has reversed operands (shift amount first)
-  auto shr = V_LSHRREV_B32::create(builder, loc, vregType, *rhs, *lhs);
-  ctx.getMapper().mapValue(shrOp.getResult(), shr);
-  return success();
-}
-
-/// Handle arith.shrsi - emit v_ashrrev_i32 (signed right shift)
-LogicalResult handleArithShRSI(Operation *op, TranslationContext &ctx) {
-  auto shrOp = cast<arith::ShRSIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(shrOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(shrOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  // v_ashrrev has reversed operands (shift amount first)
-  auto shr = V_ASHRREV_I32::create(builder, loc, vregType, *rhs, *lhs);
-  ctx.getMapper().mapValue(shrOp.getResult(), shr);
-  return success();
-}
-
-/// Handle arith.extui - emit v_mov (zero extension is implicit)
-LogicalResult handleArithExtUI(Operation *op, TranslationContext &ctx) {
-  auto extOp = cast<arith::ExtUIOp>(op);
-
-  // Zero extension to 32-bit is usually a no-op on GPU
-  auto src = ctx.getMapper().getMapped(extOp.getIn());
-  if (src) {
-    ctx.getMapper().mapValue(extOp.getResult(), *src);
-  }
-  return success();
-}
-
-/// Handle arith.extsi - emit v_bfe (bit field extract for sign extension)
-LogicalResult handleArithExtSI(Operation *op, TranslationContext &ctx) {
-  auto extOp = cast<arith::ExtSIOp>(op);
-
-  // Sign extension - for small types, may need bit field extract
-  // For now, pass through (assuming 32-bit operations)
-  auto src = ctx.getMapper().getMapped(extOp.getIn());
-  if (src) {
-    ctx.getMapper().mapValue(extOp.getResult(), *src);
-  }
-  return success();
-}
-
-/// Handle arith.trunci - emit v_and (truncation via mask)
-LogicalResult handleArithTruncI(Operation *op, TranslationContext &ctx) {
-  auto truncOp = cast<arith::TruncIOp>(op);
-
-  // Truncation is usually a no-op (just use lower bits)
-  auto src = ctx.getMapper().getMapped(truncOp.getIn());
-  if (src) {
-    ctx.getMapper().mapValue(truncOp.getResult(), *src);
-  }
-  return success();
-}
-
-/// Handle arith.cmpi - emit v_cmp instructions (sets VCC implicitly)
-LogicalResult handleArithCmpI(Operation *op, TranslationContext &ctx) {
-  auto cmpOp = cast<arith::CmpIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  auto lhs = ctx.getMapper().getMapped(cmpOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(cmpOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  // Emit comparison based on predicate
-  // These operations set VCC implicitly (no SSA result)
-  switch (cmpOp.getPredicate()) {
-  case arith::CmpIPredicate::eq:
-    V_CMP_EQ_U32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::ne:
-    V_CMP_NE_U32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::slt:
-    V_CMP_LT_I32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::sle:
-    V_CMP_LE_I32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::sgt:
-    V_CMP_GT_I32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::sge:
-    V_CMP_GE_I32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::ult:
-    V_CMP_LT_U32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::ule:
-    V_CMP_LE_U32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::ugt:
-    V_CMP_GT_U32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpIPredicate::uge:
-    V_CMP_GE_U32::create(builder, loc, *lhs, *rhs);
-    break;
-  }
-
-  // For now, we don't track VCC as a value - the select handler uses it
-  // implicitly Map result to a placeholder (the select handler will use VCC)
-  auto immOne = ctx.createImmType(1);
-  auto one = ConstantOp::create(builder, loc, immOne, 1);
-  ctx.getMapper().mapValue(cmpOp.getResult(), one);
-
-  return success();
-}
-
-/// Handle arith.select - emit v_cndmask_b32
-LogicalResult handleArithSelect(Operation *op, TranslationContext &ctx) {
-  auto selectOp = cast<arith::SelectOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  // Note: condition was handled by cmpi which set VCC
-  auto trueVal = ctx.getMapper().getMapped(selectOp.getTrueValue());
-  auto falseVal = ctx.getMapper().getMapped(selectOp.getFalseValue());
-
-  if (!trueVal || !falseVal) {
-    return op->emitError("operands not mapped");
-  }
-
-  // Get a zero constant for the VCC operand (VCC is implicitly used)
-  auto immZero = ctx.createImmType(0);
-  auto zeroConst = ConstantOp::create(builder, loc, immZero, 0);
-
-  // v_cndmask_b32: select true_val if vcc set, false_val otherwise
-  // The third operand is VCC (implicitly used from previous v_cmp)
-  auto select = V_CNDMASK_B32::create(builder, loc, vregType, *falseVal,
-                                      *trueVal, zeroConst);
-  ctx.getMapper().mapValue(selectOp.getResult(), select);
-  return success();
-}
-
-/// Helper to check if an MLIR value is a constant power of 2
-static std::optional<int64_t> getConstantPowerOf2(Value val,
-                                                  TranslationContext &ctx) {
-  // Check if it's a waveasm.constant
-  if (auto constOp = val.getDefiningOp<ConstantOp>()) {
-    int64_t value = constOp.getValue();
-    if (value > 0 && (value & (value - 1)) == 0) {
-      // It's a power of 2, return log2
-      int64_t log2val = 0;
-      while ((1LL << log2val) < value) {
-        log2val++;
-      }
-      return log2val;
-    }
-  }
-  // Check for arith.constant in the original MLIR
-  auto *defOp = val.getDefiningOp();
-  if (!defOp)
-    return std::nullopt;
-
-  if (auto arithConst = dyn_cast<arith::ConstantOp>(defOp)) {
-    if (auto intAttr = dyn_cast<IntegerAttr>(arithConst.getValue())) {
-      int64_t value = intAttr.getInt();
-      if (value > 0 && (value & (value - 1)) == 0) {
-        int64_t log2val = 0;
-        while ((1LL << log2val) < value) {
-          log2val++;
-        }
-        return log2val;
-      }
-    }
-  }
-  return std::nullopt;
-}
-
-/// Handle arith.divui - emit v_lshrrev for power-of-2, general div otherwise
-LogicalResult handleArithDivUI(Operation *op, TranslationContext &ctx) {
-  auto divOp = cast<arith::DivUIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(divOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(divOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  // Check if RHS is a constant power of 2
-  if (auto log2val = getConstantPowerOf2(divOp.getRhs(), ctx)) {
-    // Division by 2^n = right shift by n
-    auto shiftAmt = ctx.createImmType(*log2val);
-    auto shiftConst = ConstantOp::create(builder, loc, shiftAmt, *log2val);
-    auto div = V_LSHRREV_B32::create(builder, loc, vregType, shiftConst, *lhs);
-    ctx.getMapper().mapValue(divOp.getResult(), div);
-    return success();
-  }
-
-  // General case: need complex reciprocal sequence
-  // For now, emit a placeholder shift (will need full implementation)
-  // TODO: Implement proper integer division using reciprocal approximation
-  auto div = V_LSHRREV_B32::create(builder, loc, vregType, *rhs, *lhs);
-  ctx.getMapper().mapValue(divOp.getResult(), div);
-  return success();
-}
-
-/// Helper to get constant value from arith.constant
-static std::optional<int64_t> getConstantValue(Value val) {
-  auto *defOp = val.getDefiningOp();
-  if (!defOp)
-    return std::nullopt;
-
-  // Try typed arith::ConstantOp first
-  if (auto arithConst = dyn_cast<arith::ConstantOp>(defOp)) {
-    if (auto intAttr = dyn_cast<IntegerAttr>(arithConst.getValue())) {
-      return intAttr.getInt();
-    }
-  }
-
-  // Handle generic form "arith.constant" with attribute "value"
-  if (defOp->getName().getStringRef() == "arith.constant") {
-    if (auto valueAttr = defOp->getAttr("value")) {
-      if (auto intAttr = dyn_cast<IntegerAttr>(valueAttr)) {
-        return intAttr.getInt();
-      }
-    }
-  }
-
-  return std::nullopt;
-}
-
-/// Handle arith.remui - emit v_and for power-of-2, general rem otherwise
-LogicalResult handleArithRemUI(Operation *op, TranslationContext &ctx) {
-  auto remOp = cast<arith::RemUIOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(remOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(remOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  // Check if RHS is a constant power of 2
-  if (auto rhsVal = getConstantValue(remOp.getRhs())) {
-    int64_t value = *rhsVal;
-    if (value > 0 && (value & (value - 1)) == 0) {
-      // Modulo by 2^n = AND with (2^n - 1)
-      int64_t mask = value - 1;
-      auto maskImm = ctx.createImmType(mask);
-      auto maskConst = ConstantOp::create(builder, loc, maskImm, mask);
-      auto rem = V_AND_B32::create(builder, loc, vregType, *lhs, maskConst);
-      ctx.getMapper().mapValue(remOp.getResult(), rem);
-      return success();
-    }
-  }
-
-  // General case: rem = lhs - (lhs / rhs) * rhs
-  // This requires division, so for now emit a placeholder
-  // TODO: Implement proper general modulo
-  auto rem = V_AND_B32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(remOp.getResult(), rem);
-  return success();
-}
-
-//===----------------------------------------------------------------------===//
-// Floating-Point Arith Dialect Handlers
-//===----------------------------------------------------------------------===//
-
-/// Handle arith.addf - emit v_add_f32
-LogicalResult handleArithAddF(Operation *op, TranslationContext &ctx) {
-  auto addOp = cast<arith::AddFOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(addOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(addOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto add = V_ADD_F32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(addOp.getResult(), add);
-  return success();
-}
-
-/// Handle arith.subf - emit v_sub_f32
-LogicalResult handleArithSubF(Operation *op, TranslationContext &ctx) {
-  auto subOp = cast<arith::SubFOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(subOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(subOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto sub = V_SUB_F32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(subOp.getResult(), sub);
-  return success();
-}
-
-/// Handle arith.mulf - emit v_mul_f32
-LogicalResult handleArithMulF(Operation *op, TranslationContext &ctx) {
-  auto mulOp = cast<arith::MulFOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(mulOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(mulOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto mul = V_MUL_F32::create(builder, loc, vregType, *lhs, *rhs);
-  ctx.getMapper().mapValue(mulOp.getResult(), mul);
-  return success();
-}
-
-/// Handle arith.divf - emit v_div (reciprocal + multiply for fast path)
-LogicalResult handleArithDivF(Operation *op, TranslationContext &ctx) {
-  auto divOp = cast<arith::DivFOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto lhs = ctx.getMapper().getMapped(divOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(divOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  // Use v_rcp_f32 + v_mul_f32 for fast division
-  auto rcp = V_RCP_F32::create(builder, loc, vregType, *rhs);
-  auto div = V_MUL_F32::create(builder, loc, vregType, *lhs, rcp);
-  ctx.getMapper().mapValue(divOp.getResult(), div);
-  return success();
-}
-
-/// Handle arith.negf - emit v_mul with -1.0
-LogicalResult handleArithNegF(Operation *op, TranslationContext &ctx) {
-  auto negOp = cast<arith::NegFOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto operand = ctx.getMapper().getMapped(negOp.getOperand());
-  if (!operand) {
-    return op->emitError("operand not mapped");
-  }
-
-  // Negate using XOR with sign bit (0x80000000)
-  int32_t signBit = 0x80000000;
-  auto signImm = ctx.createImmType(signBit);
-  auto signConst = ConstantOp::create(builder, loc, signImm, signBit);
-  auto neg = V_XOR_B32::create(builder, loc, vregType, *operand, signConst);
-  ctx.getMapper().mapValue(negOp.getResult(), neg);
-  return success();
-}
-
-/// Handle arith.cmpf - emit v_cmp_f32 instructions
-LogicalResult handleArithCmpF(Operation *op, TranslationContext &ctx) {
-  auto cmpOp = cast<arith::CmpFOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  auto lhs = ctx.getMapper().getMapped(cmpOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(cmpOp.getRhs());
-
-  if (!lhs || !rhs) {
-    return op->emitError("operands not mapped");
-  }
-
-  // Emit comparison based on predicate (sets VCC)
-  switch (cmpOp.getPredicate()) {
-  case arith::CmpFPredicate::OEQ:
-  case arith::CmpFPredicate::UEQ:
-    V_CMP_EQ_F32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpFPredicate::ONE:
-  case arith::CmpFPredicate::UNE:
-    V_CMP_NE_F32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpFPredicate::OLT:
-  case arith::CmpFPredicate::ULT:
-    V_CMP_LT_F32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpFPredicate::OLE:
-  case arith::CmpFPredicate::ULE:
-    V_CMP_LE_F32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpFPredicate::OGT:
-  case arith::CmpFPredicate::UGT:
-    V_CMP_GT_F32::create(builder, loc, *lhs, *rhs);
-    break;
-  case arith::CmpFPredicate::OGE:
-  case arith::CmpFPredicate::UGE:
-    V_CMP_GE_F32::create(builder, loc, *lhs, *rhs);
-    break;
-  default:
-    // Handle other predicates as unordered
-    V_CMP_NE_F32::create(builder, loc, *lhs, *rhs);
-    break;
-  }
-
-  // Map result to a placeholder (select handler uses VCC implicitly)
-  auto immOne = ctx.createImmType(1);
-  auto one = ConstantOp::create(builder, loc, immOne, 1);
-  ctx.getMapper().mapValue(cmpOp.getResult(), one);
-
-  return success();
-}
-
-/// Handle math.fma - emit v_fma_f32
-LogicalResult handleMathFma(Operation *op, TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  if (op->getNumOperands() < 3) {
-    return op->emitError("fma requires 3 operands");
-  }
-
-  auto a = ctx.getMapper().getMapped(op->getOperand(0));
-  auto b = ctx.getMapper().getMapped(op->getOperand(1));
-  auto c = ctx.getMapper().getMapped(op->getOperand(2));
-
-  if (!a || !b || !c) {
-    return op->emitError("operands not mapped");
-  }
-
-  auto fma = V_FMA_F32::create(builder, loc, vregType, *a, *b, *c);
-  ctx.getMapper().mapValue(op->getResult(0), fma);
-  return success();
-}
-
-//===----------------------------------------------------------------------===//
-// Vector Dialect - Additional Handlers
-//===----------------------------------------------------------------------===//
-
-/// Handle vector.broadcast - replicate scalar to vector
-LogicalResult handleVectorBroadcast(Operation *op, TranslationContext &ctx) {
-  auto broadcastOp = cast<vector::BroadcastOp>(op);
-
-  // For GPU, broadcast is typically a no-op (value is already lane-uniform
-  // or will be handled by register allocation)
-  auto src = ctx.getMapper().getMapped(broadcastOp.getSource());
-  if (src) {
-    ctx.getMapper().mapValue(broadcastOp.getResult(), *src);
-  }
-  return success();
-}
-
-/// Handle vector.extract - extract scalar from vector
-/// Creates a new single-element register at the correct offset
-LogicalResult handleVectorExtract(Operation *op, TranslationContext &ctx) {
-  auto extractOp = cast<vector::ExtractOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  auto src = ctx.getMapper().getMapped(extractOp.getSource());
-  if (!src) {
-    return op->emitError("source value not mapped");
-  }
-
-  // Get the extraction index (position in the source vector)
-  auto staticPos = extractOp.getStaticPosition();
-  int64_t index = 0;
-  if (!staticPos.empty()) {
-    index = staticPos[0];
-  }
-
-  // Get the source register type to find the base physical register
-  Type srcType = src->getType();
-  int64_t baseIdx = 0;
-
-  if (auto pvreg = dyn_cast<PVRegType>(srcType)) {
-    // Physical VGPR - extract element at offset
-    baseIdx = pvreg.getIndex() + index;
-    auto elemType = PVRegType::get(builder.getContext(), baseIdx, 1);
-    auto elemReg = PrecoloredVRegOp::create(builder, loc, elemType, baseIdx, 1);
-    ctx.getMapper().mapValue(extractOp.getResult(), elemReg);
-  } else {
-    // Virtual VGPR or other type - use waveasm.extract op
-    // This will be lowered to proper register offset during register allocation
-    auto elemType = ctx.createVRegType(1, 1);
-    auto extractWaveOp = ExtractOp::create(builder, loc, elemType, *src,
-                                           builder.getI64IntegerAttr(index));
-    ctx.getMapper().mapValue(extractOp.getResult(), extractWaveOp.getResult());
-  }
-  return success();
-}
-
-/// Handle vector.insert - insert scalar into vector
-LogicalResult handleVectorInsert(Operation *op, TranslationContext &ctx) {
-  auto insertOp = cast<vector::InsertOp>(op);
-
-  // Pass through the destination (modification happens via register offset)
-  auto dest = ctx.getMapper().getMapped(insertOp.getDest());
-  if (dest) {
-    ctx.getMapper().mapValue(insertOp.getResult(), *dest);
-  }
-  return success();
-}
-
-/// Handle vector.shape_cast - reinterpret vector shape
-LogicalResult handleVectorShapeCast(Operation *op, TranslationContext &ctx) {
-  auto castOp = cast<vector::ShapeCastOp>(op);
-
-  // Shape cast is a no-op at the register level
-  auto src = ctx.getMapper().getMapped(castOp.getSource());
-  if (src) {
-    ctx.getMapper().mapValue(castOp.getResult(), *src);
-  }
-  return success();
-}
+// Note: GPU handlers have been moved to handlers/GPUHandlers.cpp
+// Note: Arith handlers have been moved to handlers/ArithHandlers.cpp
+// Note: Math handlers have been moved to handlers/MathHandlers.cpp
+
+// Note: Simple vector handlers (handleVectorBroadcast, handleVectorExtract,
+// handleVectorInsert, handleVectorShapeCast, handleVectorFma,
+// handleVectorReduction, handleVectorExtractStridedSlice) moved to
+// handlers/VectorHandlers.cpp
 
 /// Handle vector.transfer_read - similar to vector.load
 LogicalResult handleVectorTransferRead(Operation *op, TranslationContext &ctx) {
@@ -1625,605 +593,6 @@ LogicalResult handleVectorTransferWrite(Operation *op,
   return success();
 }
 
-// Note: gpu.subgroup_broadcast handler removed - not available in this MLIR
-// version When available, it would emit v_readlane_b32 or v_readfirstlane_b32
-
-/// Handle vector.fma - fused multiply-add
-LogicalResult handleVectorFma(Operation *op, TranslationContext &ctx) {
-  auto fmaOp = cast<vector::FMAOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  auto lhs = ctx.getMapper().getMapped(fmaOp.getLhs());
-  auto rhs = ctx.getMapper().getMapped(fmaOp.getRhs());
-  auto acc = ctx.getMapper().getMapped(fmaOp.getAcc());
-
-  if (!lhs || !rhs || !acc) {
-    return op->emitError("FMA operands not mapped");
-  }
-
-  auto resultType = fmaOp.getResult().getType();
-  Type elemType;
-  int64_t numElements = 1;
-  if (auto vecType = dyn_cast<VectorType>(resultType)) {
-    elemType = vecType.getElementType();
-    numElements = vecType.getNumElements();
-  } else {
-    elemType = resultType;
-  }
-
-  // Create result register
-  auto vregType = ctx.createVRegType(numElements, 1);
-
-  Value result;
-  if (elemType.isF32()) {
-    // v_fma_f32 dst, src0, src1, src2 : dst = src0 * src1 + src2
-    result = V_FMA_F32::create(builder, loc, vregType, *lhs, *rhs, *acc);
-  } else if (elemType.isF16()) {
-    // v_fma_f16 for f16 types
-    result = V_FMA_F16::create(builder, loc, vregType, *lhs, *rhs, *acc);
-  } else {
-    // Fall back to mul + add for other types
-    auto mulResult = V_MUL_F32::create(builder, loc, vregType, *lhs, *rhs);
-    result = V_ADD_F32::create(builder, loc, vregType, mulResult, *acc);
-  }
-
-  ctx.getMapper().mapValue(fmaOp.getResult(), result);
-  return success();
-}
-
-/// Handle vector.reduction - reduction operations
-LogicalResult handleVectorReduction(Operation *op, TranslationContext &ctx) {
-  // vector.reduction has operands: vector to reduce, optional accumulator
-  if (op->getNumOperands() < 1) {
-    return op->emitError("reduction requires vector operand");
-  }
-
-  Value vector = op->getOperand(0);
-  auto vectorMapped = ctx.getMapper().getMapped(vector);
-  if (!vectorMapped) {
-    return op->emitError("vector operand not mapped");
-  }
-
-  // For now, emit a comment - full reduction requires wave-level operations
-  // like DPP or permute instructions
-  ctx.emitComment("vector.reduction - wave-level reduction");
-
-  // Simple fallback: just map the first element
-  ctx.getMapper().mapValue(op->getResult(0), *vectorMapped);
-  return success();
-}
-
-/// Handle scf.if - emit conditional execution
-LogicalResult handleSCFIf(Operation *op, TranslationContext &ctx) {
-  auto ifOp = cast<scf::IfOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Get condition
-  auto cond = ctx.getMapper().getMapped(ifOp.getCondition());
-  if (!cond) {
-    return op->emitError("condition not mapped");
-  }
-
-  // Create labels for if/else/endif
-  std::string baseName =
-      "L_if_" + std::to_string(reinterpret_cast<uintptr_t>(op));
-  std::string elseLabel = baseName + "_else";
-  std::string endLabel = baseName + "_end";
-
-  // Branch to else block if condition is false
-  auto elseLabelRef = SymbolRefAttr::get(builder.getContext(), elseLabel);
-  S_CBRANCH_SCC0::create(builder, loc, elseLabelRef);
-
-  // Translate then region
-  for (Operation &thenOp : ifOp.getThenRegion().front().without_terminator()) {
-    if (failed(translateOperation(&thenOp, ctx))) {
-      return failure();
-    }
-  }
-
-  // Jump to end (skip else block)
-  if (!ifOp.getElseRegion().empty()) {
-    auto endLabelRef = SymbolRefAttr::get(builder.getContext(), endLabel);
-    S_BRANCH::create(builder, loc, endLabelRef);
-  }
-
-  // Else label and block
-  if (!ifOp.getElseRegion().empty()) {
-    LabelOp::create(builder, loc, elseLabel);
-    for (Operation &elseOp :
-         ifOp.getElseRegion().front().without_terminator()) {
-      if (failed(translateOperation(&elseOp, ctx))) {
-        return failure();
-      }
-    }
-  }
-
-  // End label
-  LabelOp::create(builder, loc, endLabel);
-
-  return success();
-}
-
-/// Handle memref.subview - compute offset
-LogicalResult handleMemRefSubView(Operation *op, TranslationContext &ctx) {
-  auto subviewOp = cast<memref::SubViewOp>(op);
-
-  // Pass through source (offset computation handled by indices)
-  if (auto src = ctx.getMapper().getMapped(subviewOp.getSource())) {
-    ctx.getMapper().mapValue(subviewOp.getResult(), *src);
-  }
-  return success();
-}
-
-/// Handle memref.load - emit ds_read or buffer_load
-LogicalResult handleMemRefLoad(Operation *op, TranslationContext &ctx) {
-  auto loadOp = cast<memref::LoadOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  auto memrefType = loadOp.getMemRefType();
-  auto vregType = ctx.createVRegType();
-
-  if (isLDSMemRef(memrefType)) {
-    // LDS load
-    Value vaddr;
-    if (!loadOp.getIndices().empty()) {
-      if (auto mapped = ctx.getMapper().getMapped(loadOp.getIndices()[0])) {
-        vaddr = *mapped;
-      }
-    }
-    if (!vaddr) {
-      auto immType = ctx.createImmType(0);
-      vaddr = ConstantOp::create(builder, loc, immType, 0);
-    }
-
-    auto readOp = DS_READ_B32::create(builder, loc, TypeRange{vregType}, vaddr);
-    ctx.getMapper().mapValue(loadOp.getResult(), readOp.getResult(0));
-  } else {
-    // Global load
-    auto sregType = ctx.createSRegType(4, 4);
-    auto srd = PrecoloredSRegOp::create(builder, loc, sregType, 8, 4);
-
-    Value voffset;
-    if (!loadOp.getIndices().empty()) {
-      if (auto mapped = ctx.getMapper().getMapped(loadOp.getIndices()[0])) {
-        voffset = *mapped;
-      }
-    }
-    if (!voffset) {
-      auto immType = ctx.createImmType(0);
-      voffset = ConstantOp::create(builder, loc, immType, 0);
-    }
-
-    auto loadInstr = BUFFER_LOAD_DWORD::create(
-        builder, loc, TypeRange{vregType}, srd, voffset);
-    ctx.getMapper().mapValue(loadOp.getResult(), loadInstr.getResult(0));
-  }
-
-  return success();
-}
-
-/// Handle memref.store - emit ds_write or buffer_store
-LogicalResult handleMemRefStore(Operation *op, TranslationContext &ctx) {
-  auto storeOp = cast<memref::StoreOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  auto memrefType = storeOp.getMemRefType();
-
-  auto data = ctx.getMapper().getMapped(storeOp.getValueToStore());
-  if (!data) {
-    return op->emitError("data value not mapped");
-  }
-
-  if (isLDSMemRef(memrefType)) {
-    // LDS store
-    Value vaddr;
-    if (!storeOp.getIndices().empty()) {
-      if (auto mapped = ctx.getMapper().getMapped(storeOp.getIndices()[0])) {
-        vaddr = *mapped;
-      }
-    }
-    if (!vaddr) {
-      auto immType = ctx.createImmType(0);
-      vaddr = ConstantOp::create(builder, loc, immType, 0);
-    }
-
-    DS_WRITE_B32::create(builder, loc, *data, vaddr);
-  } else {
-    // Global store
-    auto sregType = ctx.createSRegType(4, 4);
-    auto srd = PrecoloredSRegOp::create(builder, loc, sregType, 8, 4);
-
-    Value voffset;
-    if (!storeOp.getIndices().empty()) {
-      if (auto mapped = ctx.getMapper().getMapped(storeOp.getIndices()[0])) {
-        voffset = *mapped;
-      }
-    }
-    if (!voffset) {
-      auto immType = ctx.createImmType(0);
-      voffset = ConstantOp::create(builder, loc, immType, 0);
-    }
-
-    BUFFER_STORE_DWORD::create(builder, loc, *data, srd, voffset);
-  }
-
-  return success();
-}
-
-/// Handle memref.cast - pass through source
-LogicalResult handleMemRefCast(Operation *op, TranslationContext &ctx) {
-  auto castOp = cast<memref::CastOp>(op);
-
-  if (auto src = ctx.getMapper().getMapped(castOp.getSource())) {
-    ctx.getMapper().mapValue(castOp.getResult(), *src);
-  }
-  return success();
-}
-
-//===----------------------------------------------------------------------===//
-// Affine Dialect Handlers
-//===----------------------------------------------------------------------===//
-
-/// Handle affine.apply - compile affine expression to arithmetic instructions
-LogicalResult handleAffineApply(Operation *op, TranslationContext &ctx) {
-  auto applyOp = cast<affine::AffineApplyOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto vregType = ctx.createVRegType();
-
-  auto map = applyOp.getAffineMap();
-
-  // Get the single operand (for single-dimension maps)
-  if (applyOp.getOperands().empty()) {
-    return op->emitError("affine.apply with no operands");
-  }
-
-  Value baseValue;
-  if (auto mapped = ctx.getMapper().getMapped(applyOp.getOperands()[0])) {
-    baseValue = *mapped;
-  } else {
-    return op->emitError("operand not mapped");
-  }
-
-  // For single result affine maps, analyze the expression
-  if (map.getNumResults() != 1) {
-    return op->emitError("only single-result affine maps supported");
-  }
-
-  AffineExpr expr = map.getResult(0);
-
-  // Get thread ID upper bound for the first operand (used for simplification)
-  // If the first operand is a thread ID with known upper bound, we can
-  // simplify floor divisions where divisor >= upper_bound to 0
-  int64_t threadIdUpperBound = 0;
-  if (applyOp.getOperands().size() > 0) {
-    threadIdUpperBound = ctx.getThreadIdUpperBound(applyOp.getOperands()[0]);
-  }
-
-  // HIGH-LEVEL SIMPLIFICATION: Check if the entire expression simplifies to
-  // just the input symbol when floor divisions evaluate to 0 Pattern: s0 + (s0
-  // floordiv N) * C where N >= upper_bound
-  //       => s0 + 0 * C = s0
-  if (threadIdUpperBound > 0) {
-    // Check if expression is Add(symbol, Mul(FloorDiv(symbol, N), C))
-    // where N >= threadIdUpperBound
-    if (auto addExpr = dyn_cast<AffineBinaryOpExpr>(expr)) {
-      if (addExpr.getKind() == AffineExprKind::Add) {
-        // Check if LHS is the symbol and RHS is a Mul containing FloorDiv
-        if (isa<AffineSymbolExpr>(addExpr.getLHS())) {
-          if (auto mulExpr = dyn_cast<AffineBinaryOpExpr>(addExpr.getRHS())) {
-            if (mulExpr.getKind() == AffineExprKind::Mul) {
-              // Check if LHS of Mul is FloorDiv with divisor >= upperBound
-              if (auto floorExpr =
-                      dyn_cast<AffineBinaryOpExpr>(mulExpr.getLHS())) {
-                if (floorExpr.getKind() == AffineExprKind::FloorDiv) {
-                  if (auto constDiv =
-                          dyn_cast<AffineConstantExpr>(floorExpr.getRHS())) {
-                    if (constDiv.getValue() >= threadIdUpperBound) {
-                      // Expression simplifies to just the symbol (s0)
-                      // Map result to the thread ID value
-                      ctx.getMapper().mapValue(applyOp.getResult(), baseValue);
-                      return success();
-                    }
-                  }
-                }
-              }
-              // Also check RHS of Mul
-              if (auto floorExpr =
-                      dyn_cast<AffineBinaryOpExpr>(mulExpr.getRHS())) {
-                if (floorExpr.getKind() == AffineExprKind::FloorDiv) {
-                  if (auto constDiv =
-                          dyn_cast<AffineConstantExpr>(floorExpr.getRHS())) {
-                    if (constDiv.getValue() >= threadIdUpperBound) {
-                      ctx.getMapper().mapValue(applyOp.getResult(), baseValue);
-                      return success();
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-
-  // NOTE: We used to extract constant addends for buffer store offset:N
-  // optimization but this caused bugs when the affine result was used in arith
-  // operations (the constant was lost). For now, just compile the full
-  // expression.
-  // TODO: Re-enable constant extraction only for values used directly in memory
-  // ops
-  int64_t constAddend = 0;
-  AffineExpr exprToCompile = expr;
-
-  // Simple pattern matching for common affine expressions
-  // Pattern: d0 mod N -> v_and_b32 (when N is power of 2)
-  // Pattern: d0 floordiv N -> v_lshrrev_b32 (when N is power of 2)
-  // Pattern: d0 * N -> v_lshlrev_b32 (when N is power of 2)
-
-  // Result type that includes bit range tracking for OR optimization
-  struct ExprResult {
-    Value value;
-    BitRange range;
-    ExprResult(Value v, BitRange r) : value(v), range(r) {}
-  };
-
-  // Helper to emit the compiled expression with bit range tracking
-  std::function<ExprResult(AffineExpr)> compileExpr =
-      [&](AffineExpr e) -> ExprResult {
-    // Dimension reference
-    if (auto dimExpr = dyn_cast<AffineDimExpr>(e)) {
-      if (dimExpr.getPosition() < applyOp.getOperands().size()) {
-        Value operand = applyOp.getOperands()[dimExpr.getPosition()];
-        if (auto mapped = ctx.getMapper().getMapped(operand)) {
-          // Use tracked bit range if available
-          BitRange range = ctx.getBitRange(*mapped);
-          return ExprResult(*mapped, range);
-        }
-      }
-      return ExprResult(baseValue, ctx.getBitRange(baseValue));
-    }
-
-    // Symbol reference
-    if (auto symExpr = dyn_cast<AffineSymbolExpr>(e)) {
-      int64_t symIdx = map.getNumDims() + symExpr.getPosition();
-      if (symIdx < static_cast<int64_t>(applyOp.getOperands().size())) {
-        Value operand = applyOp.getOperands()[symIdx];
-        if (auto mapped = ctx.getMapper().getMapped(operand)) {
-          BitRange range = ctx.getBitRange(*mapped);
-          return ExprResult(*mapped, range);
-        }
-      }
-      return ExprResult(baseValue, ctx.getBitRange(baseValue));
-    }
-
-    // Constant
-    if (auto constExpr = dyn_cast<AffineConstantExpr>(e)) {
-      int64_t val = constExpr.getValue();
-      auto immType = ctx.createImmType(val);
-      Value constVal = ConstantOp::create(builder, loc, immType, val);
-      return ExprResult(constVal, BitRange::fromConstant(val));
-    }
-
-    // Binary expressions
-    if (auto binExpr = dyn_cast<AffineBinaryOpExpr>(e)) {
-      ExprResult lhsResult = compileExpr(binExpr.getLHS());
-      ExprResult rhsResult = compileExpr(binExpr.getRHS());
-      Value lhs = lhsResult.value;
-      Value rhs = rhsResult.value;
-      BitRange lhsRange = lhsResult.range;
-      BitRange rhsRange = rhsResult.range;
-
-      switch (binExpr.getKind()) {
-      case AffineExprKind::Add: {
-        // Check if bit ranges overlap - if not, use OR instead of ADD
-        if (!lhsRange.overlaps(rhsRange)) {
-          // Check if either operand is a shift (Mul by power of 2)
-          // If so, emit v_lshl_or_b32 directly instead of lshlrev + or
-          auto tryFuseShiftOr =
-              [&](AffineExpr shiftExpr, Value orend,
-                  BitRange orendRange) -> std::optional<ExprResult> {
-            if (auto mulExpr = dyn_cast<AffineBinaryOpExpr>(shiftExpr)) {
-              if (mulExpr.getKind() == AffineExprKind::Mul) {
-                // Check for power of 2 multiplier
-                if (auto constRhs =
-                        dyn_cast<AffineConstantExpr>(mulExpr.getRHS())) {
-                  int64_t val = constRhs.getValue();
-                  if (val > 0 && (val & (val - 1)) == 0) {
-                    // It's a shift! Emit v_lshl_or_b32 directly
-                    int64_t shiftAmount = log2(val);
-                    // Get the base value being shifted (compile without the
-                    // multiply)
-                    ExprResult baseResult = compileExpr(mulExpr.getLHS());
-                    auto shiftImm = ctx.createImmType(shiftAmount);
-                    auto shiftConst =
-                        ConstantOp::create(builder, loc, shiftImm, shiftAmount);
-                    // v_lshl_or_b32: dst = (src << shift) | orend
-                    Value fusedResult = V_LSHL_OR_B32::create(
-                        builder, loc, vregType, baseResult.value, shiftConst,
-                        orend);
-                    BitRange shiftedRange =
-                        baseResult.range.shiftLeft(shiftAmount);
-                    BitRange resultRange = shiftedRange.merge(orendRange);
-                    ctx.setBitRange(fusedResult, resultRange);
-                    return ExprResult(fusedResult, resultRange);
-                  }
-                }
-                // Also check LHS for constant
-                if (auto constLhs =
-                        dyn_cast<AffineConstantExpr>(mulExpr.getLHS())) {
-                  int64_t val = constLhs.getValue();
-                  if (val > 0 && (val & (val - 1)) == 0) {
-                    int64_t shiftAmount = log2(val);
-                    ExprResult baseResult = compileExpr(mulExpr.getRHS());
-                    auto shiftImm = ctx.createImmType(shiftAmount);
-                    auto shiftConst =
-                        ConstantOp::create(builder, loc, shiftImm, shiftAmount);
-                    Value fusedResult = V_LSHL_OR_B32::create(
-                        builder, loc, vregType, baseResult.value, shiftConst,
-                        orend);
-                    BitRange shiftedRange =
-                        baseResult.range.shiftLeft(shiftAmount);
-                    BitRange resultRange = shiftedRange.merge(orendRange);
-                    ctx.setBitRange(fusedResult, resultRange);
-                    return ExprResult(fusedResult, resultRange);
-                  }
-                }
-              }
-            }
-            return std::nullopt;
-          };
-
-          // Try to fuse: check if LHS is a shift
-          if (auto result = tryFuseShiftOr(binExpr.getLHS(), rhs, rhsRange)) {
-            return *result;
-          }
-          // Try to fuse: check if RHS is a shift
-          if (auto result = tryFuseShiftOr(binExpr.getRHS(), lhs, lhsRange)) {
-            return *result;
-          }
-
-          // No fusion possible, emit regular v_or_b32
-          Value orResult = V_OR_B32::create(builder, loc, vregType, lhs, rhs);
-          BitRange resultRange = lhsRange.merge(rhsRange);
-          ctx.setBitRange(orResult, resultRange);
-          return ExprResult(orResult, resultRange);
-        }
-        // Overlapping ranges - must use ADD
-        Value addResult = V_ADD_U32::create(builder, loc, vregType, lhs, rhs);
-        BitRange resultRange = lhsRange.extendForAdd(rhsRange);
-        ctx.setBitRange(addResult, resultRange);
-        return ExprResult(addResult, resultRange);
-      }
-
-      case AffineExprKind::Mul: {
-        // Constant folding: if either operand is constant 0, result is 0
-        if (auto constLhs = dyn_cast<AffineConstantExpr>(binExpr.getLHS())) {
-          if (constLhs.getValue() == 0) {
-            auto immZero = ctx.createImmType(0);
-            return ExprResult(ConstantOp::create(builder, loc, immZero, 0),
-                              BitRange(0, 0));
-          }
-        }
-        if (auto constRhs = dyn_cast<AffineConstantExpr>(binExpr.getRHS())) {
-          if (constRhs.getValue() == 0) {
-            auto immZero = ctx.createImmType(0);
-            return ExprResult(ConstantOp::create(builder, loc, immZero, 0),
-                              BitRange(0, 0));
-          }
-          // Check if RHS is constant power of 2 -> use shift
-          int64_t val = constRhs.getValue();
-          if (isPowerOf2(val)) {
-            int64_t shiftAmount = log2(val);
-            auto shiftAmt = ctx.createImmType(shiftAmount);
-            auto shiftConst =
-                ConstantOp::create(builder, loc, shiftAmt, shiftAmount);
-            Value shiftResult =
-                V_LSHLREV_B32::create(builder, loc, vregType, shiftConst, lhs);
-            // Shift the bit range left by shiftAmount
-            BitRange resultRange = lhsRange.shiftLeft(shiftAmount);
-            ctx.setBitRange(shiftResult, resultRange);
-            return ExprResult(shiftResult, resultRange);
-          }
-        }
-        // Also check LHS for power of 2 multiply
-        if (auto constLhs = dyn_cast<AffineConstantExpr>(binExpr.getLHS())) {
-          int64_t val = constLhs.getValue();
-          if (isPowerOf2(val)) {
-            int64_t shiftAmount = log2(val);
-            auto shiftAmt = ctx.createImmType(shiftAmount);
-            auto shiftConst =
-                ConstantOp::create(builder, loc, shiftAmt, shiftAmount);
-            Value shiftResult =
-                V_LSHLREV_B32::create(builder, loc, vregType, shiftConst, rhs);
-            BitRange resultRange = rhsRange.shiftLeft(shiftAmount);
-            ctx.setBitRange(shiftResult, resultRange);
-            return ExprResult(shiftResult, resultRange);
-          }
-        }
-        Value mulResult =
-            V_MUL_LO_U32::create(builder, loc, vregType, lhs, rhs);
-        return ExprResult(mulResult, BitRange()); // Conservative: full range
-      }
-
-      case AffineExprKind::FloorDiv: {
-        // Check if RHS is constant
-        if (auto constRhs = dyn_cast<AffineConstantExpr>(binExpr.getRHS())) {
-          int64_t divisor = constRhs.getValue();
-
-          // SIMPLIFICATION: If the LHS is a thread ID with upper_bound <=
-          // divisor, then floor(tid / divisor) = 0 for all valid thread IDs.
-          // Example: tid_x in [0, 63], floor(tid_x / 64) = 0
-          if (threadIdUpperBound > 0 && divisor >= threadIdUpperBound) {
-            // LHS is in range [0, upper_bound-1], so floor(LHS / divisor) = 0
-            auto immZero = ctx.createImmType(0);
-            return ExprResult(ConstantOp::create(builder, loc, immZero, 0),
-                              BitRange(0, 0));
-          }
-
-          // Check if divisor is power of 2 -> use right shift
-          if (isPowerOf2(divisor)) {
-            int64_t shiftAmount = log2(divisor);
-            auto shiftAmt = ctx.createImmType(shiftAmount);
-            auto shiftConst =
-                ConstantOp::create(builder, loc, shiftAmt, shiftAmount);
-            Value shiftResult =
-                V_LSHRREV_B32::create(builder, loc, vregType, shiftConst, lhs);
-            // Shift the bit range right by shiftAmount
-            BitRange resultRange = lhsRange.shiftRight(shiftAmount);
-            ctx.setBitRange(shiftResult, resultRange);
-            return ExprResult(shiftResult, resultRange);
-          }
-        }
-        // General floordiv - needs more complex handling
-        return ExprResult(lhs, BitRange()); // Conservative
-      }
-
-      case AffineExprKind::Mod: {
-        // Check if RHS is constant power of 2 -> use AND
-        if (auto constRhs = dyn_cast<AffineConstantExpr>(binExpr.getRHS())) {
-          int64_t val = constRhs.getValue();
-          if (isPowerOf2(val)) {
-            auto maskVal = ctx.createImmType(val - 1);
-            auto maskConst = ConstantOp::create(builder, loc, maskVal, val - 1);
-            Value andResult =
-                V_AND_B32::create(builder, loc, vregType, lhs, maskConst);
-            // Result uses bits 0..(log2(val)-1)
-            BitRange resultRange = BitRange(0, log2(val) - 1);
-            ctx.setBitRange(andResult, resultRange);
-            return ExprResult(andResult, resultRange);
-          }
-        }
-        // General mod - needs more complex handling
-        return ExprResult(lhs, BitRange()); // Conservative
-      }
-
-      default:
-        return ExprResult(lhs,
-                          BitRange()); // Unsupported, return LHS as fallback
-      }
-    }
-
-    return ExprResult(baseValue, BitRange()); // Fallback
-  };
-
-  ExprResult result = compileExpr(exprToCompile);
-  ctx.getMapper().mapValue(applyOp.getResult(), result.value);
-  ctx.setBitRange(result.value, result.range);
-
-  // Track the constant addend for buffer store offset:N optimization
-  if (constAddend != 0) {
-    ctx.setConstOffset(applyOp.getResult(), constAddend);
-  }
-
-  return success();
-}
-
 //===----------------------------------------------------------------------===//
 // Vector Dialect Handlers
 //===----------------------------------------------------------------------===//
@@ -2239,6 +608,12 @@ LogicalResult handleVectorLoad(Operation *op, TranslationContext &ctx) {
   auto memrefType = cast<MemRefType>(loadOp.getBase().getType());
   auto vectorType = loadOp.getVectorType();
   int64_t numBytes = getVectorBytes(vectorType);
+
+  // Validate vector size - zero-size vectors would cause division issues
+  if (numBytes <= 0) {
+    return op->emitError("vector.load with zero-size vector type is invalid");
+  }
+
   int64_t numDwords = (numBytes + 3) / 4;
 
   if (isLDSMemRef(memrefType)) {
@@ -2531,6 +906,12 @@ LogicalResult handleVectorStore(Operation *op, TranslationContext &ctx) {
   auto memrefType = cast<MemRefType>(storeOp.getBase().getType());
   auto vectorType = storeOp.getVectorType();
   int64_t numBytes = getVectorBytes(vectorType);
+
+  // Validate vector size - zero-size vectors would cause issues
+  if (numBytes <= 0) {
+    return op->emitError("vector.store with zero-size vector type is invalid");
+  }
+
   [[maybe_unused]] int64_t numDwords = (numBytes + 3) / 4;
 
   auto data = ctx.getMapper().getMapped(storeOp.getValueToStore());
@@ -2785,508 +1166,50 @@ LogicalResult handleVectorStore(Operation *op, TranslationContext &ctx) {
   return success();
 }
 
-/// Handle vector.extract_strided_slice - extract subset of source registers
-/// Creates a register alias at the correct offset for proper assembly emission
-LogicalResult handleVectorExtractStridedSlice(Operation *op,
-                                              TranslationContext &ctx) {
-  auto extractOp = cast<vector::ExtractStridedSliceOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  auto src = ctx.getMapper().getMapped(extractOp.getSource());
-  if (!src) {
-    return op->emitError("source value not mapped");
-  }
-
-  // Get the offset from the offsets attribute (for 1D, it's a single value)
-  auto offsets = extractOp.getOffsets();
-  int64_t offset = 0;
-  if (!offsets.empty()) {
-    offset = cast<IntegerAttr>(offsets[0]).getInt();
-  }
-
-  // Get the size from the sizes attribute (for 1D, it's a single value)
-  auto sizes = extractOp.getSizes();
-  int64_t size = 1;
-  if (!sizes.empty()) {
-    size = cast<IntegerAttr>(sizes[0]).getInt();
-  }
-
-  // Get the source register type to find the base physical register
-  Type srcType = src->getType();
-
-  if (auto pvreg = dyn_cast<PVRegType>(srcType)) {
-    // Physical VGPR - extract element(s) at offset
-    int64_t baseIdx = pvreg.getIndex() + offset;
-    auto elemType = PVRegType::get(builder.getContext(), baseIdx, size);
-    auto elemReg =
-        PrecoloredVRegOp::create(builder, loc, elemType, baseIdx, size);
-    ctx.getMapper().mapValue(extractOp.getResult(), elemReg);
-  } else {
-    // Virtual VGPR or other type - use waveasm.extract op
-    // This will be lowered to proper register offset during register allocation
-    auto elemType = ctx.createVRegType(size, 1);
-    auto extractWaveOp = ExtractOp::create(builder, loc, elemType, *src,
-                                           builder.getI64IntegerAttr(offset));
-    ctx.getMapper().mapValue(extractOp.getResult(), extractWaveOp.getResult());
-  }
-  return success();
-}
-
 //===----------------------------------------------------------------------===//
 // AMDGPU Dialect Handlers
 //===----------------------------------------------------------------------===//
 
 /// Handle amdgpu.lds_barrier - emit s_waitcnt + s_barrier
-LogicalResult handleAMDGPULdsBarrier(Operation *op, TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Wait for all LDS operations to complete
-  S_WAITCNT_LGKMCNT::create(builder, loc, 0);
-  // Barrier synchronization
-  S_BARRIER::create(builder, loc);
-
-  return success();
-}
-
-/// Handle amdgpu.mfma - emit v_mfma instruction
-/// Supports all MFMA variants: f16, bf16, i8, f32, f64, fp8/bf8
-LogicalResult handleAMDGPUMfma(Operation *op, TranslationContext &ctx) {
-  auto mfmaOp = cast<amdgpu::MFMAOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Get MFMA dimensions
-  int64_t m = mfmaOp.getM();
-  int64_t n = mfmaOp.getN();
-  int64_t k = mfmaOp.getK();
-
-  // Get operands
-  auto srcA = ctx.getMapper().getMapped(mfmaOp.getSourceA());
-  auto srcB = ctx.getMapper().getMapped(mfmaOp.getSourceB());
-  auto srcC = ctx.getMapper().getMapped(mfmaOp.getDestC());
-
-  if (!srcA || !srcB || !srcC) {
-    return op->emitError("MFMA operands not mapped");
-  }
-
-  // Determine element type from sourceA
-  Type srcAType = mfmaOp.getSourceA().getType();
-  Type elemType;
-  if (auto vecType = dyn_cast<VectorType>(srcAType)) {
-    elemType = vecType.getElementType();
-  } else {
-    elemType = srcAType;
-  }
-
-  // Determine accumulator size based on MFMA variant
-  int64_t accSize = 4; // Default for 16x16xk small variants
-  if (m == 32 && n == 32) {
-    accSize = 16; // 32x32xk variants
-  } else if (m == 4 && n == 4) {
-    accSize = 4; // 4x4xk variants
-  } else if (m == 16 && n == 16 && k == 4) {
-    accSize = 16; // 16x16x4 variants (larger accumulator)
-  }
-
-  // For f64 variants, sizes are different
-  if (elemType.isF64()) {
-    if (m == 16 && n == 16)
-      accSize = 8;
-    else if (m == 4 && n == 4)
-      accSize = 2;
-  }
-
-  auto vregType = ctx.createVRegType(accSize, 4); // Quad-aligned
-
-  // Select MFMA instruction based on dimensions and element type
-  Value result;
-
-  // Helper to check if element type matches
-  auto isF16 = [&]() { return elemType.isF16(); };
-  auto isBF16 = [&]() { return elemType.isBF16(); };
-  auto isI8 = [&]() { return elemType.isInteger(8); };
-  auto isF32 = [&]() { return elemType.isF32(); };
-  auto isF64 = [&]() { return elemType.isF64(); };
-
-  // F16 variants
-  if (isF16()) {
-    if (m == 16 && n == 16 && k == 16) {
-      result = V_MFMA_F32_16X16X16_F16::create(builder, loc, vregType, *srcA,
-                                               *srcB, *srcC);
-    } else if (m == 16 && n == 16 && k == 32) {
-      // gfx950+ (MI350) - v_mfma_f32_16x16x32_f16
-      result = V_MFMA_F32_16X16X32_F16::create(builder, loc, vregType, *srcA,
-                                               *srcB, *srcC);
-    } else if (m == 32 && n == 32 && k == 8) {
-      result = V_MFMA_F32_32X32X8_F16::create(builder, loc, vregType, *srcA,
-                                              *srcB, *srcC);
-    } else if (m == 16 && n == 16 && k == 4) {
-      auto largeVregType = ctx.createVRegType(16, 4);
-      result = V_MFMA_F32_16X16X4_F16::create(builder, loc, largeVregType,
-                                              *srcA, *srcB, *srcC);
-    } else if (m == 32 && n == 32 && k == 4) {
-      auto largeVregType = ctx.createVRegType(32, 4);
-      result = V_MFMA_F32_32X32X4_F16::create(builder, loc, largeVregType,
-                                              *srcA, *srcB, *srcC);
-    } else if (m == 4 && n == 4 && k == 4) {
-      result = V_MFMA_F32_4X4X4_F16::create(builder, loc, vregType, *srcA,
-                                            *srcB, *srcC);
-    } else {
-      result = V_MFMA_F32_16X16X16_F16::create(builder, loc, vregType, *srcA,
-                                               *srcB, *srcC);
-    }
-  }
-  // BF16 variants
-  else if (isBF16()) {
-    if (m == 16 && n == 16 && k == 16) {
-      result = V_MFMA_F32_16X16X16_BF16::create(builder, loc, vregType, *srcA,
-                                                *srcB, *srcC);
-    } else if (m == 16 && n == 16 && k == 32) {
-      // gfx950+ (MI350) - v_mfma_f32_16x16x32_bf16
-      result = V_MFMA_F32_16X16X32_BF16::create(builder, loc, vregType, *srcA,
-                                                *srcB, *srcC);
-    } else if (m == 32 && n == 32 && k == 8) {
-      result = V_MFMA_F32_32X32X8_BF16::create(builder, loc, vregType, *srcA,
-                                               *srcB, *srcC);
-    } else if (m == 16 && n == 16 && k == 4) {
-      auto largeVregType = ctx.createVRegType(16, 4);
-      result = V_MFMA_F32_16X16X4_BF16::create(builder, loc, largeVregType,
-                                               *srcA, *srcB, *srcC);
-    } else if (m == 32 && n == 32 && k == 4) {
-      auto largeVregType = ctx.createVRegType(32, 4);
-      result = V_MFMA_F32_32X32X4_BF16::create(builder, loc, largeVregType,
-                                               *srcA, *srcB, *srcC);
-    } else if (m == 4 && n == 4 && k == 4) {
-      result = V_MFMA_F32_4X4X4_BF16::create(builder, loc, vregType, *srcA,
-                                             *srcB, *srcC);
-    } else {
-      result = V_MFMA_F32_16X16X16_BF16::create(builder, loc, vregType, *srcA,
-                                                *srcB, *srcC);
-    }
-  }
-  // I8 variants
-  else if (isI8()) {
-    if (m == 16 && n == 16 && k == 16) {
-      result = V_MFMA_I32_16X16X16_I8::create(builder, loc, vregType, *srcA,
-                                              *srcB, *srcC);
-    } else if (m == 32 && n == 32 && k == 8) {
-      result = V_MFMA_I32_32X32X8_I8::create(builder, loc, vregType, *srcA,
-                                             *srcB, *srcC);
-    } else if (m == 16 && n == 16 && k == 4) {
-      auto largeVregType = ctx.createVRegType(16, 4);
-      result = V_MFMA_I32_16X16X4_I8::create(builder, loc, largeVregType, *srcA,
-                                             *srcB, *srcC);
-    } else if (m == 32 && n == 32 && k == 4) {
-      auto largeVregType = ctx.createVRegType(32, 4);
-      result = V_MFMA_I32_32X32X4_I8::create(builder, loc, largeVregType, *srcA,
-                                             *srcB, *srcC);
-    } else if (m == 4 && n == 4 && k == 4) {
-      result = V_MFMA_I32_4X4X4_I8::create(builder, loc, vregType, *srcA, *srcB,
-                                           *srcC);
-    } else {
-      result = V_MFMA_I32_16X16X16_I8::create(builder, loc, vregType, *srcA,
-                                              *srcB, *srcC);
-    }
-  }
-  // F32 variants
-  else if (isF32()) {
-    if (m == 16 && n == 16 && k == 4) {
-      result = V_MFMA_F32_16X16X4_F32::create(builder, loc, vregType, *srcA,
-                                              *srcB, *srcC);
-    } else if (m == 32 && n == 32 && k == 2) {
-      result = V_MFMA_F32_32X32X2_F32::create(builder, loc, vregType, *srcA,
-                                              *srcB, *srcC);
-    } else if (m == 4 && n == 4 && k == 1) {
-      result = V_MFMA_F32_4X4X1_F32::create(builder, loc, vregType, *srcA,
-                                            *srcB, *srcC);
-    } else {
-      result = V_MFMA_F32_16X16X4_F32::create(builder, loc, vregType, *srcA,
-                                              *srcB, *srcC);
-    }
-  }
-  // F64 variants
-  else if (isF64()) {
-    if (m == 16 && n == 16 && k == 4) {
-      auto f64VregType = ctx.createVRegType(8, 4);
-      result = V_MFMA_F64_16X16X4_F64::create(builder, loc, f64VregType, *srcA,
-                                              *srcB, *srcC);
-    } else if (m == 4 && n == 4 && k == 4) {
-      auto f64VregType = ctx.createVRegType(2, 2);
-      result = V_MFMA_F64_4X4X4_F64::create(builder, loc, f64VregType, *srcA,
-                                            *srcB, *srcC);
-    } else {
-      auto f64VregType = ctx.createVRegType(8, 4);
-      result = V_MFMA_F64_16X16X4_F64::create(builder, loc, f64VregType, *srcA,
-                                              *srcB, *srcC);
-    }
-  }
-  // Default to F16 16x16x16
-  else {
-    result = V_MFMA_F32_16X16X16_F16::create(builder, loc, vregType, *srcA,
-                                             *srcB, *srcC);
-  }
-
-  ctx.getMapper().mapValue(mfmaOp.getDestD(), result);
-  return success();
-}
-
-//===----------------------------------------------------------------------===//
-// MemRef Dialect Handlers
-//===----------------------------------------------------------------------===//
-
-/// Handle memref.alloc - track LDS allocation
-LogicalResult handleMemRefAlloc(Operation *op, TranslationContext &ctx) {
-  auto allocOp = cast<memref::AllocOp>(op);
-
-  // Check if this is an LDS allocation (workgroup address space)
-  auto memrefType = allocOp.getResult().getType();
-  if (isLDSMemRef(memrefType)) {
-    // Compute the allocation size in bytes
-    int64_t numElements = 1;
-    for (int64_t dim : memrefType.getShape()) {
-      if (dim != ShapedType::kDynamic) {
-        numElements *= dim;
-      }
-    }
-    int64_t elementBytes = (memrefType.getElementTypeBitWidth() + 7) / 8;
-    int64_t allocSize = numElements * elementBytes;
-
-    // Track the total LDS size for the kernel descriptor
-    ctx.addLDSSize(allocSize);
-  }
-
-  return success();
-}
-
-/// Handle memref.view - compute LDS offset
-LogicalResult handleMemRefView(Operation *op, TranslationContext &ctx) {
-  auto viewOp = cast<memref::ViewOp>(op);
-
-  // Track the byte offset for LDS addressing
-  // The byteShift is the base offset into LDS that this view starts at
-  if (auto offset = ctx.getMapper().getMapped(viewOp.getByteShift())) {
-    // Store the LDS base offset for this memref so it can be added
-    // during vector.load/store operations
-    ctx.setLDSBaseOffset(viewOp.getResult(), *offset);
-  }
-
-  return success();
-}
-
-/// Handle memref.reinterpret_cast - track memref identity
-LogicalResult handleMemRefReinterpretCast(Operation *op,
-                                          TranslationContext &ctx) {
-  auto castOp = cast<memref::ReinterpretCastOp>(op);
-
-  // Reinterpret cast doesn't change the underlying buffer
-  if (auto src = ctx.getMapper().getMapped(castOp.getSource())) {
-    ctx.getMapper().mapValue(castOp.getResult(), *src);
-  }
-
-  // Propagate LDS base offset from source to result
-  // This is needed when memref.view creates an LDS view with a byte offset,
-  // and then memref.reinterpret_cast is applied to reshape it
-  if (auto ldsOffset = ctx.getLDSBaseOffset(castOp.getSource())) {
-    ctx.setLDSBaseOffset(castOp.getResult(), *ldsOffset);
-  }
-
-  // The result type often has more specific shape info than the source.
-  // Update the SRD buffer size if the result type is larger.
-  if (auto memrefType = dyn_cast<MemRefType>(castOp.getResult().getType())) {
-    int64_t bufferSize = computeBufferSizeFromMemRef(memrefType);
-    // Update the SRD for the source memref with the more accurate size
-    ctx.updateSRDBufferSize(castOp.getSource(), bufferSize);
-  }
-
-  return success();
-}
-
 //===----------------------------------------------------------------------===//
 // SCF Dialect Handlers
 //===----------------------------------------------------------------------===//
 
+/// Handle scf.for - emit loop structure with iter_args support
+} // namespace
+
+namespace waveasm {
+//===----------------------------------------------------------------------===//
+// Forward declarations for handlers in external files
+//===----------------------------------------------------------------------===//
+
+// From AffineHandlers.cpp
+LogicalResult handleAffineApply(Operation *op, TranslationContext &ctx);
+
+// From MemRefHandlers.cpp
+LogicalResult handleMemRefAlloc(Operation *op, TranslationContext &ctx);
+LogicalResult handleMemRefView(Operation *op, TranslationContext &ctx);
+LogicalResult handleMemRefReinterpretCast(Operation *op,
+                                          TranslationContext &ctx);
+LogicalResult handleMemRefSubView(Operation *op, TranslationContext &ctx);
+LogicalResult handleMemRefLoad(Operation *op, TranslationContext &ctx);
+LogicalResult handleMemRefStore(Operation *op, TranslationContext &ctx);
+LogicalResult handleMemRefCast(Operation *op, TranslationContext &ctx);
+
+// From SCFHandlers.cpp
 LogicalResult handleSCFFor(Operation *op, TranslationContext &ctx);
 LogicalResult handleSCFIf(Operation *op, TranslationContext &ctx);
 LogicalResult handleSCFYield(Operation *op, TranslationContext &ctx);
 
-/// Handle scf.for - emit loop structure with iter_args support
-LogicalResult handleSCFFor(Operation *op, TranslationContext &ctx) {
-  auto forOp = cast<scf::ForOp>(op);
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Generate unique loop label
-  std::string labelName = ctx.generateLabel("L_loop");
-  std::string endLabelName = labelName + "_end";
-
-  // Initialize loop counter in a FIXED physical SGPR
-  // Using a fixed register ensures the counter value persists across loop
-  // iterations Reserve s32+ for loop counters (avoiding s0-s31 which may be
-  // used for arguments/SRDs/cache-swizzle) Note: s[24:31] may be used for cache
-  // swizzle SRDs in g2s kernels
-  int64_t loopCounterPhysReg =
-      32 + ctx.getLoopDepth(); // Use s32, s33, etc. for nested loops
-  auto counterType =
-      PSRegType::get(builder.getContext(), loopCounterPhysReg, 1);
-
-  auto lb = ctx.getMapper().getMapped(forOp.getLowerBound());
-  if (!lb) {
-    auto immType = ctx.createImmType(0);
-    lb = ConstantOp::create(builder, loc, immType, 0);
-  }
-
-  auto counter = S_MOV_B32::create(builder, loc, counterType, *lb);
-  ctx.getMapper().mapValue(forOp.getInductionVar(), counter);
-
-  // Handle iter_args (loop-carried values)
-  // These are values passed from one iteration to the next (e.g., accumulators)
-  // For vector-type iter_args (accumulators), we need to:
-  // 1. Allocate VREGs before the loop
-  // 2. Initialize them with v_mov_b32
-  // 3. Map the region arg to those VREGs for in-place accumulation
-  SmallVector<Value, 4> iterArgValues;
-  for (auto [initArg, regionArg] :
-       llvm::zip(forOp.getInitArgs(), forOp.getRegionIterArgs())) {
-    auto mapped = ctx.getMapper().getMapped(initArg);
-
-    // Check if this is a vector type (likely an accumulator for MFMA)
-    Type regionArgType = regionArg.getType();
-    if (auto vecType = dyn_cast<VectorType>(regionArgType)) {
-      // For vector accumulators, we need to materialize VREGs
-      int64_t numElems = vecType.getNumElements();
-
-      // Check if the init value is a constant 0 (common for accumulators)
-      bool isZeroInit = false;
-      if (auto constOp = initArg.getDefiningOp<arith::ConstantOp>()) {
-        if (auto denseAttr = dyn_cast<DenseElementsAttr>(constOp.getValue())) {
-          if (denseAttr.isSplat()) {
-            if (auto floatAttr =
-                    dyn_cast<FloatAttr>(denseAttr.getSplatValue<Attribute>())) {
-              isZeroInit = (floatAttr.getValueAsDouble() == 0.0);
-            }
-          }
-        }
-      }
-
-      if (isZeroInit) {
-        // Allocate VREGs for the accumulator and initialize to 0
-        // The MFMA instruction will use these registers for in-place
-        // accumulation
-        auto vregType =
-            ctx.createVRegType(numElems, 4); // Quad-aligned for MFMA
-        auto zeroImm = ctx.createImmType(0);
-        auto zero = ConstantOp::create(builder, loc, zeroImm, 0);
-
-        // Create v_mov_b32 to initialize each element of the accumulator
-        // Mark with "no_cse" to prevent CSE from merging multiple accumulators
-        auto accMovOp = V_MOV_B32::create(builder, loc, vregType, zero);
-        accMovOp->setAttr("no_cse", builder.getUnitAttr());
-        Value accReg = accMovOp.getResult();
-
-        // Map the region argument to this register
-        ctx.getMapper().mapValue(regionArg, accReg);
-        iterArgValues.push_back(accReg);
-        continue;
-      }
-    }
-
-    // Default: map to the initial value directly
-    if (mapped) {
-      ctx.getMapper().mapValue(regionArg, *mapped);
-      iterArgValues.push_back(*mapped);
-    }
-  }
-
-  // Set up loop context for nested operations
-  LoopContext loopCtx;
-  loopCtx.inductionVar = counter;
-  loopCtx.iterArgs = iterArgValues;
-  loopCtx.labelName = labelName;
-  loopCtx.depth = ctx.getLoopDepth() + 1;
-  ctx.pushLoopContext(loopCtx);
-
-  // Clear expression cache at loop entry (loop-variant expressions must be
-  // recomputed)
-  ctx.clearExprCache();
-
-  // Loop label
-  LabelOp::create(builder, loc, labelName);
-
-  // Translate loop body
-  for (Operation &bodyOp : forOp.getBody()->without_terminator()) {
-    if (failed(translateOperation(&bodyOp, ctx))) {
-      ctx.popLoopContext();
-      return failure();
-    }
-  }
-
-  // Handle yield - update iter_args for next iteration
-  if (auto yieldOp = dyn_cast<scf::YieldOp>(forOp.getBody()->getTerminator())) {
-    for (auto [yieldedVal, regionArg] :
-         llvm::zip(yieldOp.getOperands(), forOp.getRegionIterArgs())) {
-      auto mapped = ctx.getMapper().getMapped(yieldedVal);
-      if (mapped) {
-        // Update the region argument mapping for the next iteration
-        ctx.getMapper().mapValue(regionArg, *mapped);
-      }
-    }
-  }
-
-  // Increment counter - use the same physical register type as the counter
-  // This ensures the increment writes to the same physical register
-  auto step = ctx.getMapper().getMapped(forOp.getStep());
-  if (!step) {
-    auto immType = ctx.createImmType(1);
-    step = ConstantOp::create(builder, loc, immType, 1);
-  }
-
-  // Get the counter type from the original counter (it's a physical register
-  // type)
-  auto counterPhysType = counter.getType();
-  auto newCounter =
-      S_ADD_U32::create(builder, loc, counterPhysType, counter, *step);
-
-  // Since we're using physical registers, newCounter is in the same register as
-  // counter Update the mapping for any post-loop uses
-  ctx.getMapper().mapValue(forOp.getInductionVar(), newCounter);
-
-  // Compare and branch back to loop header
-  auto ub = ctx.getMapper().getMapped(forOp.getUpperBound());
-  if (ub) {
-    S_CMP_LT_U32::create(builder, loc, newCounter, *ub);
-    auto labelRef = SymbolRefAttr::get(builder.getContext(), labelName);
-    S_CBRANCH_SCC1::create(builder, loc, labelRef);
-  }
-
-  // End label for loop exit
-  LabelOp::create(builder, loc, endLabelName);
-
-  // Map loop results to final iter_arg values
-  if (auto yieldOp = dyn_cast<scf::YieldOp>(forOp.getBody()->getTerminator())) {
-    for (auto [result, yieldedVal] :
-         llvm::zip(forOp.getResults(), yieldOp.getOperands())) {
-      auto mapped = ctx.getMapper().getMapped(yieldedVal);
-      if (mapped) {
-        ctx.getMapper().mapValue(result, *mapped);
-      }
-    }
-  }
-
-  ctx.popLoopContext();
-  return success();
-}
-
-/// Handle scf.yield - typically a no-op
-LogicalResult handleSCFYield(Operation *op, TranslationContext &ctx) {
-  // Yield values are handled by the parent op
-  return success();
-}
-
-} // namespace
-
-namespace waveasm {
+// From AMDGPUHandlers.cpp
+LogicalResult handleAMDGPULdsBarrier(Operation *op, TranslationContext &ctx);
+LogicalResult handleAMDGPUMfma(Operation *op, TranslationContext &ctx);
+LogicalResult handleFatRawBufferCast(Operation *op, TranslationContext &ctx);
+LogicalResult handleGatherToLds(Operation *op, TranslationContext &ctx);
+LogicalResult handleRawBufferLoad(Operation *op, TranslationContext &ctx);
+LogicalResult handleRawBufferStore(Operation *op, TranslationContext &ctx);
+LogicalResult handleReadFirstLane(Operation *op, TranslationContext &ctx);
+LogicalResult handleSWaitcnt(Operation *op, TranslationContext &ctx);
 
 //===----------------------------------------------------------------------===//
 // OpHandlerRegistry Implementation
@@ -3349,627 +1272,6 @@ LogicalResult handleBindingSubspan(Operation *op, TranslationContext &ctx) {
   // Queue SRD setup for this binding
   ctx.queueSRDSetup(op->getResult(0), bindingIdx, bufferSize);
 
-  return success();
-}
-
-/// Handle amdgpu.fat_raw_buffer_cast
-/// This operation creates a buffer descriptor with cache swizzle info for
-/// gather_to_lds
-LogicalResult handleFatRawBufferCast(Operation *op, TranslationContext &ctx) {
-  // The operation format is:
-  //   %result = amdgpu.fat_raw_buffer_cast %source, %offset,
-  //   %cacheSwizzleStride
-  //
-  // For gather_to_lds, we need to create a new SRD with cache swizzle bits:
-  //   word0: copy from source SRD word0 (base address low)
-  //   word1: (source word1 & 0xffff) | 0x40400000 (cache swizzle bits)
-  //   word2: 0x7ffffffd (size for gather)
-  //   word3: 0x27000 (format with swizzle, instead of standard 0x20000)
-
-  if (op->getNumOperands() < 1) {
-    return success(); // No source operand
-  }
-
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Get source SRD
-  auto srcMapped = ctx.getMapper().getMapped(op->getOperand(0));
-  if (!srcMapped) {
-    // Just pass through if no mapping
-    return success();
-  }
-
-  // Check if there's a cache swizzle stride (operand 2)
-  bool hasCacheSwizzle = false;
-  int64_t swizzleStride = 0;
-  if (op->getNumOperands() >= 3) {
-    if (auto swizzleVal = getConstantValue(op->getOperand(2))) {
-      swizzleStride = *swizzleVal;
-      hasCacheSwizzle = (swizzleStride > 0);
-    }
-  }
-
-  if (!hasCacheSwizzle) {
-    // No cache swizzle - just pass through
-    ctx.getMapper().mapValue(op->getResult(0), *srcMapped);
-    return success();
-  }
-
-  // Create a new SRD with cache swizzle bits
-  // Allocate a new SRD slot (use SGPRs after the regular SRDs)
-  // The Python backend uses s[24:27] for the first swizzle SRD
-  int64_t newSrdBase = ctx.getNextSwizzleSRDIndex();
-
-  // Get source SRD base index
-  int64_t srcSrdBase = -1;
-
-  // Check if srcMapped is defined by a PrecoloredSRegOp (which has the physical
-  // index)
-  if (auto defOp = srcMapped->getDefiningOp()) {
-    if (defOp->getName().getStringRef() == "waveasm.precolored.sreg") {
-      if (auto indexAttr = defOp->getAttrOfType<IntegerAttr>("index")) {
-        srcSrdBase = indexAttr.getInt();
-      }
-    }
-  }
-
-  // Fallback: try PSRegType
-  if (srcSrdBase < 0) {
-    if (auto psreg = dyn_cast<PSRegType>(srcMapped->getType())) {
-      srcSrdBase = psreg.getIndex();
-    }
-  }
-
-  // Fallback: try getSRDIndex
-  if (srcSrdBase < 0) {
-    if (auto srcSrdIdx = ctx.getSRDIndex(op->getOperand(0))) {
-      srcSrdBase = *srcSrdIdx;
-    }
-  }
-
-  if (srcSrdBase < 0) {
-    // Fallback - just pass through
-    ctx.getMapper().mapValue(op->getResult(0), *srcMapped);
-    return success();
-  }
-
-  // Emit the cache swizzle SRD setup:
-  //   s_mov_b32 sN, sSrc       ; word0 - base address low
-  //   s_and_b32 sN+1, sSrc+1, 0xffff
-  //   s_or_b32 sN+1, sN+1, 0x40400000  ; cache swizzle bits
-  //   s_mov_b32 sN+2, 0x7ffffffd  ; word2 - size
-  //   s_mov_b32 sN+3, 0x27000     ; word3 - format with swizzle
-
-  // word0: copy base address low
-  std::string mov0 = "s_mov_b32 s" + std::to_string(newSrdBase) + ", s" +
-                     std::to_string(srcSrdBase);
-  RawOp::create(builder, loc, mov0);
-
-  // word1: mask and add cache swizzle bits
-  std::string and1 = "s_and_b32 s" + std::to_string(newSrdBase + 1) + ", s" +
-                     std::to_string(srcSrdBase + 1) + ", 0xffff";
-  RawOp::create(builder, loc, and1);
-
-  std::string or1 = "s_or_b32 s" + std::to_string(newSrdBase + 1) + ", s" +
-                    std::to_string(newSrdBase + 1) + ", 0x40400000";
-  RawOp::create(builder, loc, or1);
-
-  // word2: size for gather operations
-  std::string mov2 =
-      "s_mov_b32 s" + std::to_string(newSrdBase + 2) + ", 0x7ffffffd";
-  RawOp::create(builder, loc, mov2);
-
-  // word3: format with swizzle (0x27000 instead of 0x20000)
-  std::string mov3 =
-      "s_mov_b32 s" + std::to_string(newSrdBase + 3) + ", 0x27000";
-  RawOp::create(builder, loc, mov3);
-
-  // Create precolored SREG for the new SRD and map the result
-  auto srdType = ctx.createSRegType(4, 4);
-  auto newSrd = PrecoloredSRegOp::create(builder, loc, srdType, newSrdBase, 4);
-  ctx.getMapper().mapValue(op->getResult(0), newSrd);
-
-  // Also store the swizzle stride for potential later use
-  ctx.setCacheSwizzleStride(op->getResult(0), swizzleStride);
-
-  return success();
-}
-
-/// Handle amdgpu.gather_to_lds
-/// This is the high-level operation for gathering from global memory to LDS
-///
-/// Operand layout (from MLIR):
-///   operand(0) = source buffer (fat_raw_buffer memref)
-///   operand(1) = source index (global memory offset) -> voffset for
-///   buffer_load operand(2) = LDS destination memref -> used to get LDS base
-///   offset operand(3) = LDS row index -> used to compute m0 operand(4) = LDS
-///   col index -> used to compute m0
-///
-/// For buffer_load...lds:
-///   - voffset provides the global memory offset
-///   - m0 provides the LDS destination offset
-LogicalResult handleGatherToLds(Operation *op, TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // operand(0): source buffer descriptor
-  Value srcBuffer = op->getOperand(0);
-  auto srcMapped = ctx.getMapper().getMapped(srcBuffer);
-  if (!srcMapped) {
-    return op->emitError("source buffer not mapped");
-  }
-
-  // Get element size from transferType attribute for scaling index to bytes
-  // The MLIR source index is in element units, but buffer_load needs byte
-  // offset
-  int64_t elemBytes = 2; // Default f16
-  if (auto transferTypeAttr = op->getAttrOfType<TypeAttr>("transferType")) {
-    if (auto vecType = dyn_cast<VectorType>(transferTypeAttr.getValue())) {
-      auto elemType = vecType.getElementType();
-      elemBytes = elemType.getIntOrFloatBitWidth() / 8;
-    }
-  }
-
-  // operand(1): source index -> this is the voffset for global memory access
-  // IMPORTANT: Each gather_to_lds needs its OWN voffset VGPR because
-  // buffer_load_dword...lds is asynchronous. If we reuse the same VGPR
-  // for multiple gather_to_lds operations, subsequent operations may
-  // overwrite the voffset before the previous load completes.
-  // Also, the source index from MLIR is in ELEMENTS, not bytes, so we need
-  // to scale by the element size.
-  Value voff;
-  if (op->getNumOperands() > 1) {
-    Value srcIndex = op->getOperand(1);
-    auto srcIndexMapped = ctx.getMapper().getMapped(srcIndex);
-    if (srcIndexMapped) {
-      // Scale index from elements to bytes
-      auto vregType = ctx.createVRegType();
-      if (elemBytes != 1) {
-        // Use shifts for power-of-2 scaling
-        if (llvm::isPowerOf2_64(elemBytes)) {
-          int64_t shiftAmt = llvm::Log2_64(elemBytes);
-          auto shiftImm = ctx.createImmType(shiftAmt);
-          auto shiftConst =
-              ConstantOp::create(builder, loc, shiftImm, shiftAmt);
-          voff = V_LSHLREV_B32::create(builder, loc, vregType, shiftConst,
-                                       *srcIndexMapped);
-        } else {
-          // Fallback: multiply for non-power-of-2
-          auto scaleImm = ctx.createImmType(elemBytes);
-          auto scaleConst =
-              ConstantOp::create(builder, loc, scaleImm, elemBytes);
-          voff = V_MUL_LO_U32::create(builder, loc, vregType, scaleConst,
-                                      *srcIndexMapped);
-        }
-      } else {
-        // elemBytes == 1, no scaling needed, just copy to fresh VGPR
-        voff = V_MOV_B32::create(builder, loc, vregType, *srcIndexMapped);
-      }
-    }
-  }
-  if (!voff) {
-    // Fallback: use zero offset
-    auto immType = ctx.createImmType(0);
-    auto zeroConst = ConstantOp::create(builder, loc, immType, 0);
-    auto vregType = ctx.createVRegType();
-    voff = V_MOV_B32::create(builder, loc, vregType, zeroConst);
-  }
-
-  // operand(2): LDS destination memref -> get LDS base offset and shape info
-  int64_t ldsBaseOffsetConst = 0;
-  int64_t ldsRowStride = 0; // bytes per row in LDS
-  bool hasLdsBaseOffset = false;
-
-  if (op->getNumOperands() > 2) {
-    mlir::Value ldsMemref = op->getOperand(2);
-
-    // Get the LDS base offset from memref.view
-    if (auto baseOffset = ctx.getLDSBaseOffset(ldsMemref)) {
-      // Try to extract constant value from the mapped offset
-      if (auto immType = dyn_cast<ImmType>(baseOffset->getType())) {
-        ldsBaseOffsetConst = immType.getValue();
-        hasLdsBaseOffset = true;
-      }
-    }
-
-    // Get the LDS memref shape to compute row stride
-    // The LDS layout is typically [rows, cols] with element type f16 (2 bytes)
-    if (auto memrefType = dyn_cast<MemRefType>(ldsMemref.getType())) {
-      auto shape = memrefType.getShape();
-      int64_t elementBytes = 2; // Default f16
-      if (auto elemType = memrefType.getElementType()) {
-        elementBytes = elemType.getIntOrFloatBitWidth() / 8;
-      }
-      if (shape.size() >= 2) {
-        // row stride = number of columns * element bytes
-        ldsRowStride = shape[1] * elementBytes;
-      }
-    }
-  }
-
-  // Get element size for column offset calculation
-  int64_t ldsElemBytes = 2; // Default f16
-  if (op->getNumOperands() > 2) {
-    if (auto memrefType = dyn_cast<MemRefType>(op->getOperand(2).getType())) {
-      if (auto elemType = memrefType.getElementType()) {
-        ldsElemBytes = elemType.getIntOrFloatBitWidth() / 8;
-      }
-    }
-  }
-
-  // operand(3) and operand(4): LDS row and col indices -> used to compute m0
-  // M0 = lds_base + row * row_stride + col * elem_bytes
-  //    = lds_base + (row * lds_cols + col) * elem_bytes
-  int64_t m0Const = ldsBaseOffsetConst;
-  bool canUseImmediateM0 = hasLdsBaseOffset;
-  int64_t colIdxConst = 0;
-  bool hasConstCol = false;
-
-  // Get column index if available (operand 4)
-  if (op->getNumOperands() > 4) {
-    mlir::Value ldsColIndex = op->getOperand(4);
-    auto ldsColMapped = ctx.getMapper().getMapped(ldsColIndex);
-    if (ldsColMapped) {
-      if (auto immType = dyn_cast<ImmType>(ldsColMapped->getType())) {
-        colIdxConst = immType.getValue();
-        hasConstCol = true;
-      }
-    }
-  }
-
-  if (op->getNumOperands() > 3) {
-    mlir::Value ldsRowIndex = op->getOperand(3);
-    auto ldsRowMapped = ctx.getMapper().getMapped(ldsRowIndex);
-
-    if (ldsRowMapped) {
-      // Try to extract constant value from row index
-      if (auto immType = dyn_cast<ImmType>(ldsRowMapped->getType())) {
-        int64_t rowIdxConst = immType.getValue();
-        // Compute m0 = base + row * stride + col * elem_bytes
-        if (ldsRowStride > 0) {
-          m0Const = ldsBaseOffsetConst + rowIdxConst * ldsRowStride;
-          if (hasConstCol) {
-            m0Const += colIdxConst * ldsElemBytes;
-          }
-        } else {
-          // Fallback: assume row index is already in bytes
-          m0Const = ldsBaseOffsetConst + rowIdxConst;
-        }
-        canUseImmediateM0 = (op->getNumOperands() <= 4) || hasConstCol;
-      } else {
-        // Row index is not a constant, need to compute dynamically
-        canUseImmediateM0 = false;
-      }
-    }
-  }
-
-  // Set M0 for LDS destination
-  if (canUseImmediateM0) {
-    // Use immediate value directly - most efficient
-    auto m0Imm = ctx.createImmType(m0Const);
-    auto m0ConstVal = ConstantOp::create(builder, loc, m0Imm, m0Const);
-    S_MOV_B32_M0::create(builder, loc, m0ConstVal);
-  } else {
-    // Need to compute m0 dynamically (fallback path)
-    Value m0Val;
-    if (op->getNumOperands() > 3) {
-      mlir::Value ldsRowIndex = op->getOperand(3);
-      auto ldsRowMapped = ctx.getMapper().getMapped(ldsRowIndex);
-      if (ldsRowMapped) {
-        m0Val = *ldsRowMapped;
-
-        // Convert SGPR to VGPR if needed (v_mov_b32 can't take SGPR as source)
-        // Use v_add_u32 with zero to broadcast SGPR to VGPR
-        auto convertToVgpr = [&](Value val) -> Value {
-          if (!isVGPRType(val.getType())) {
-            // v_add_u32 can take SGPR as one operand and broadcasts it to all
-            // lanes
-            auto zeroImm = ctx.createImmType(0);
-            auto zeroConst = ConstantOp::create(builder, loc, zeroImm, 0);
-            Value zeroVgpr = V_MOV_B32::create(builder, loc,
-                                               ctx.createVRegType(), zeroConst);
-            return V_ADD_U32::create(builder, loc, ctx.createVRegType(),
-                                     zeroVgpr, val);
-          }
-          return val;
-        };
-
-        // Multiply row index by row stride if needed
-        if (ldsRowStride > 1) {
-          auto strideImm = ctx.createImmType(ldsRowStride);
-          auto strideConst =
-              ConstantOp::create(builder, loc, strideImm, ldsRowStride);
-          m0Val = convertToVgpr(m0Val);
-          m0Val = V_MUL_LO_U32::create(builder, loc, ctx.createVRegType(),
-                                       m0Val, strideConst);
-        }
-
-        // Add column offset (col * elem_bytes)
-        if (op->getNumOperands() > 4) {
-          mlir::Value ldsColIndex = op->getOperand(4);
-          auto ldsColMapped = ctx.getMapper().getMapped(ldsColIndex);
-          if (ldsColMapped) {
-            if (auto immType = dyn_cast<ImmType>(ldsColMapped->getType())) {
-              int64_t colOffset = immType.getValue() * ldsElemBytes;
-              if (colOffset != 0) {
-                auto colImm = ctx.createImmType(colOffset);
-                auto colConst =
-                    ConstantOp::create(builder, loc, colImm, colOffset);
-                Value colVgpr = V_MOV_B32::create(
-                    builder, loc, ctx.createVRegType(), colConst);
-                m0Val = convertToVgpr(m0Val);
-                m0Val = V_ADD_U32::create(builder, loc, ctx.createVRegType(),
-                                          m0Val, colVgpr);
-              }
-            } else {
-              // Column is dynamic - need to multiply by elem_bytes and add
-              m0Val = convertToVgpr(m0Val);
-              Value colVgpr = convertToVgpr(*ldsColMapped);
-              if (ldsElemBytes > 1) {
-                // Multiply col by elem_bytes
-                auto scaleImm = ctx.createImmType(ldsElemBytes);
-                auto scaleConst =
-                    ConstantOp::create(builder, loc, scaleImm, ldsElemBytes);
-                colVgpr = V_MUL_LO_U32::create(
-                    builder, loc, ctx.createVRegType(), colVgpr, scaleConst);
-              }
-              m0Val = V_ADD_U32::create(builder, loc, ctx.createVRegType(),
-                                        m0Val, colVgpr);
-            }
-          }
-        }
-
-        // Add base offset
-        if (hasLdsBaseOffset && ldsBaseOffsetConst != 0) {
-          auto baseImm = ctx.createImmType(ldsBaseOffsetConst);
-          auto baseConst =
-              ConstantOp::create(builder, loc, baseImm, ldsBaseOffsetConst);
-          Value baseVgpr =
-              V_MOV_B32::create(builder, loc, ctx.createVRegType(), baseConst);
-          m0Val = convertToVgpr(m0Val);
-          m0Val = V_ADD_U32::create(builder, loc, ctx.createVRegType(), m0Val,
-                                    baseVgpr);
-        }
-      }
-    }
-
-    if (m0Val) {
-      Value m0Src = m0Val;
-      // If source is a VGPR, convert to SGPR using v_readfirstlane_b32
-      if (isVGPRType(m0Src.getType())) {
-        auto sregType = ctx.createSRegType();
-        m0Src = V_READFIRSTLANE_B32::create(builder, loc, sregType, m0Src);
-      }
-      S_MOV_B32_M0::create(builder, loc, m0Src);
-    } else {
-      // Fallback: use zero LDS offset
-      auto zeroImm = ctx.createImmType(0);
-      auto zeroConst = ConstantOp::create(builder, loc, zeroImm, 0);
-      S_MOV_B32_M0::create(builder, loc, zeroConst);
-    }
-  }
-
-  // Determine transfer size from transferType attribute
-  int64_t transferBytes = 4; // Default: 4 bytes (buffer_load_dword_lds)
-  if (auto transferTypeAttr = op->getAttrOfType<TypeAttr>("transferType")) {
-    if (auto vecType = dyn_cast<VectorType>(transferTypeAttr.getValue())) {
-      int64_t numElems = vecType.getNumElements();
-      auto elemType = vecType.getElementType();
-      int64_t elemSize = elemType.getIntOrFloatBitWidth() / 8;
-      transferBytes = numElems * elemSize;
-    }
-  }
-
-  // Emit appropriate buffer_load instruction based on transfer size
-  auto soffImm = ctx.createImmType(0);
-  auto soffConst = ConstantOp::create(builder, loc, soffImm, 0);
-  if (transferBytes == 16) {
-    // 16 bytes = buffer_load_dwordx4_lds
-    BUFFER_LOAD_DWORDX4_LDS::create(builder, loc, voff, *srcMapped, soffConst);
-  } else if (transferBytes == 4) {
-    // 4 bytes = buffer_load_dword_lds (default)
-    BUFFER_LOAD_DWORD_LDS::create(builder, loc, voff, *srcMapped, soffConst);
-  } else {
-    return op->emitError("unsupported transfer size for gather_to_lds: " +
-                         std::to_string(transferBytes) + " bytes");
-  }
-
-  // Emit vmcnt wait after each gather_to_lds to ensure the voffset register
-  // is not reused before the load completes. This is necessary because
-  // buffer_load_dword_lds is asynchronous and the register allocator doesn't
-  // understand that the voffset must remain valid until the load completes.
-  // TODO: A more optimal solution would be to extend liveness analysis to
-  // keep voffset live until the next vmcnt(0) or barrier.
-  S_WAITCNT_VMCNT::create(builder, loc, /*count=*/0);
-
-  return success();
-}
-
-/// Handle amdgpu.raw_buffer_load - direct buffer load
-LogicalResult handleRawBufferLoad(Operation *op, TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Get operands: memref, indices, bounds check, indexOffset
-  if (op->getNumOperands() < 1) {
-    return op->emitError("raw_buffer_load requires at least a memref operand");
-  }
-
-  Value memref = op->getOperand(0);
-  auto srcMapped = ctx.getMapper().getMapped(memref);
-  if (!srcMapped) {
-    return op->emitError("buffer descriptor not mapped");
-  }
-
-  // Get offset operand if present
-  Value voffset;
-  if (op->getNumOperands() > 1) {
-    auto voffMapped = ctx.getMapper().getMapped(op->getOperand(1));
-    if (voffMapped) {
-      voffset = *voffMapped;
-    }
-  }
-  if (!voffset) {
-    auto immZero = ctx.createImmType(0);
-    voffset = ConstantOp::create(builder, loc, immZero, 0);
-  }
-
-  // Determine load width from result type
-  Type resultType = op->getResult(0).getType();
-  int64_t numElements = 1;
-  if (auto vecType = dyn_cast<VectorType>(resultType)) {
-    numElements = vecType.getNumElements();
-  }
-
-  Operation *loadInstr;
-  if (numElements == 1) {
-    auto vregType = ctx.createVRegType(1);
-    loadInstr = BUFFER_LOAD_DWORD::create(builder, loc, TypeRange{vregType},
-                                          *srcMapped, voffset);
-  } else if (numElements == 2) {
-    auto vregType = ctx.createVRegType(2, 2);
-    loadInstr = BUFFER_LOAD_DWORDX2::create(builder, loc, TypeRange{vregType},
-                                            *srcMapped, voffset);
-  } else if (numElements == 4) {
-    auto vregType = ctx.createVRegType(4, 4);
-    loadInstr = BUFFER_LOAD_DWORDX4::create(builder, loc, TypeRange{vregType},
-                                            *srcMapped, voffset);
-  } else {
-    return op->emitError("unsupported buffer load width");
-  }
-
-  ctx.getMapper().mapValue(op->getResult(0), loadInstr->getResult(0));
-  return success();
-}
-
-/// Handle amdgpu.raw_buffer_store - direct buffer store
-LogicalResult handleRawBufferStore(Operation *op, TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Get operands: value, memref, indices
-  if (op->getNumOperands() < 2) {
-    return op->emitError("raw_buffer_store requires value and memref operands");
-  }
-
-  Value storeValue = op->getOperand(0);
-  Value memref = op->getOperand(1);
-
-  auto valueMapped = ctx.getMapper().getMapped(storeValue);
-  auto srd = ctx.getMapper().getMapped(memref);
-
-  if (!valueMapped) {
-    return op->emitError("store value not mapped");
-  }
-  if (!srd) {
-    return op->emitError("buffer descriptor not mapped");
-  }
-
-  // Get offset operand if present
-  Value voffset;
-  if (op->getNumOperands() > 2) {
-    auto voffMapped = ctx.getMapper().getMapped(op->getOperand(2));
-    if (voffMapped) {
-      voffset = *voffMapped;
-    }
-  }
-  if (!voffset) {
-    auto immZero = ctx.createImmType(0);
-    voffset = ConstantOp::create(builder, loc, immZero, 0);
-  }
-
-  // Determine store width from value type
-  Type valueType = storeValue.getType();
-  int64_t numElements = 1;
-  if (auto vecType = dyn_cast<VectorType>(valueType)) {
-    numElements = vecType.getNumElements();
-  }
-
-  // Use signature: (data, srd, voffset)
-  if (numElements == 1) {
-    BUFFER_STORE_DWORD::create(builder, loc, *valueMapped, *srd, voffset);
-  } else if (numElements == 2) {
-    BUFFER_STORE_DWORDX2::create(builder, loc, *valueMapped, *srd, voffset);
-  } else if (numElements == 4) {
-    BUFFER_STORE_DWORDX4::create(builder, loc, *valueMapped, *srd, voffset);
-  } else {
-    return op->emitError("unsupported buffer store width");
-  }
-
-  return success();
-}
-
-/// Handle rocdl.readfirstlane
-LogicalResult handleReadFirstLane(Operation *op, TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-  auto sregType = ctx.createSRegType();
-
-  if (op->getNumOperands() < 1) {
-    return op->emitError("readfirstlane requires an operand");
-  }
-
-  auto src = ctx.getMapper().getMapped(op->getOperand(0));
-  if (!src) {
-    return op->emitError("source operand not mapped");
-  }
-
-  auto result = V_READFIRSTLANE_B32::create(builder, loc, sregType, *src);
-  ctx.getMapper().mapValue(op->getResult(0), result);
-
-  return success();
-}
-
-/// Handle rocdl.s.waitcnt
-LogicalResult handleSWaitcnt(Operation *op, TranslationContext &ctx) {
-  auto &builder = ctx.getBuilder();
-  auto loc = op->getLoc();
-
-  // Extract waitcnt bitfield from operation
-  // GFX9+ encoding:
-  //   bits 3:0 = vmcnt[3:0]
-  //   bits 6:4 = expcnt
-  //   bits 11:8 = lgkmcnt[3:0]
-  //   bits 15:14 = vmcnt[5:4] (high bits)
-  //
-  // A value of 15 (0xF) for lgkmcnt means "don't wait" (max outstanding)
-  // A value of 63 (0x3F) for vmcnt means "don't wait" (6-bit counter on GFX9+)
-  int64_t vmcnt = -1, lgkmcnt = -1; // -1 means "not specified"
-
-  if (auto bitfieldAttr = op->getAttrOfType<IntegerAttr>("bitfield")) {
-    int64_t bitfield = bitfieldAttr.getInt();
-    // Decode vmcnt (6 bits: low 4 bits + high 2 bits)
-    int64_t vmcnt_lo = bitfield & 0xF;
-    int64_t vmcnt_hi = (bitfield >> 14) & 0x3;
-    int64_t vmcnt_full = vmcnt_lo | (vmcnt_hi << 4);
-    // Decode lgkmcnt (4 bits)
-    int64_t lgkmcnt_full = (bitfield >> 8) & 0xF;
-
-    // Only emit wait if not max value (max = "don't wait")
-    if (vmcnt_full < 63) { // 63 = 0x3F = max vmcnt
-      vmcnt = vmcnt_full;
-    }
-    if (lgkmcnt_full < 15) { // 15 = 0xF = max lgkmcnt
-      lgkmcnt = lgkmcnt_full;
-    }
-  }
-
-  // If neither wait is needed, skip emitting
-  if (vmcnt < 0 && lgkmcnt < 0) {
-    return success();
-  }
-
-  // Create the appropriate waitcnt
-  if (vmcnt >= 0 && lgkmcnt >= 0) {
-    auto vmcntAttr = builder.getI32IntegerAttr(vmcnt);
-    auto lgkmcntAttr = builder.getI32IntegerAttr(lgkmcnt);
-    S_WAITCNT::create(builder, loc, vmcntAttr, lgkmcntAttr, IntegerAttr());
-  } else if (vmcnt >= 0) {
-    S_WAITCNT_VMCNT::create(builder, loc, vmcnt);
-  } else {
-    S_WAITCNT_LGKMCNT::create(builder, loc, lgkmcnt);
-  }
   return success();
 }
